### 2025-07-08 #8553 @editorialbot
[PRE REVIEW]: LogoClim: WorldClim in NetLogo
**Submitting author:** @danielvartan (<a href="http://orcid.org/0000-0001-7782-759X">Daniel Vartanian</a>)
**Repository:** https://github.com/sustentarea/logoclim
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
