### 2025-07-09 4bf6aa0
Add JOSS badge to README
