### 2025-07-09 #8553 @danielvartan
Hi @kthyng,

Thanks for all your work and the helpful information.

I checked the `editorialbot` output and everything looks good. The article has 1008 words—just slightly over the 1000-word limit—but I assume an 8-word excess won’t be a problem.

To streamline communication, I’m tagging the other authors here: @leandromtg @alinenutri.

I’m currently searching the JOSS reviewer database for good matches and have already identified a few reviewers with expertise in Agent-Based Models. I’ll share my suggestions soon.

@leandromtg and @alinenutri, if you have any reviewer suggestions, feel free to mention them here—just don’t tag their GitHub handles (if they have one).
