### 2025-07-14 0179729
Add `plot-max-y-range`
