### 2025-07-21 #8553 @danielvartan
Hi @kthyng,

We’ve now finalized our reviewer suggestions. Based on the JOSS reviewer database and additional research, we’d like to suggest the following reviewers:

From the JOSS reviewer database:

1. Emily Dolson ([github](https://github.com/emilydolson))
2. James Millington ([github](https://github.com/jamesdamillington))
3.  Brennan Klein ([github](https://github.com/jkbren))

External reviewers (if possible):

1. Carsten Lemmen ([github](https://github.com/platipodium)) – Researcher at the Hereon Institute (Germany), with expertise in NetLogo and environmental agent-based modeling. He authored the only NetLogo-based model [currently published](https://joss.theoj.org/papers/10.21105/joss.05731) in JOSS.
2. Eulàlia Baulenas ([email](mailto:eulalia.baulenas@gmail.com)) – PhD in Forest and Environmental Policy, with relevant experience in socio-ecological systems modeling using NetLogo.

We have no personal or professional connection with any of the suggested reviewers. There is no [conflict of interest](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy).

Please let me know if you need any additional information or if we should revise the list in any way.
