### 2025-07-23 #8553 @danielvartan
The creation of a testing suite is currently under discussion among NetLogo developers. There have been previous efforts (e.g., the [Test Extension](https://github.com/NetLogo/Test-Extension)), but these projects have since been abandoned. [Steve Railsback](https://scholar.google.com/citations?user=sCekWDEAAAAJ&hl=en&oi=ao), a prominent figure in NetLogo and computational science, recently proposed some ideas on the NetLogo forum:

https://forum.netlogo.org/t/code-testing-tools/174/4

That said, as I mentioned before, this discussion mostly concerns models with varying parameters and stochastic behavior. LogoClim is a submodel—designed to be used alongside other models—that brings WorldClim data into NetLogo, functioning more like an API package.
