### 2025-07-29 3c62ed5
Change license from MIT to GPLv3
