### 2025-07-29 #8553 @kthyng
@danielvartan I don't know netlogo much and I am going to have to trust you have these items covered as you imply, though of course items will come up in review as appropriate.
