### 2025-07-30 910c62c
Update interface images to reflect the new model version
