### 2025-07-30 f48fc0a
Update paper to align with NetLogo 7.0
