+++ README.qmd
@@ -88,7 +88,7 @@ cat(
 
 ## Overview
 
-`LogoClim` is a [NetLogo](https://ccl.northwestern.edu/netlogo/) model for simulating and visualizing global climate conditions. It allows researchers to integrate high-resolution climate data into agent-based models, supporting reproducible research in ecology, agriculture, environmental science, and other fields that rely on climate data integration.
+`LogoClim` is a [NetLogo](https://www.netlogo.org) model for simulating and visualizing global climate conditions. It allows researchers to integrate high-resolution climate data into agent-based models, supporting reproducible research in ecology, agriculture, environmental science, and other fields that rely on climate data integration.
 
 The model utilizes raster data to represent climate variables such as temperature and precipitation over time. It incorporates historical data (1951-2024) and future climate projections (2021-2100) derived from global climate models under various Shared Socioeconomic Pathways ([SSPs](https://climatedata.ca/resource/understanding-shared-socio-economic-pathways-ssps/), [O'Neill et al., 2017](https://doi.org/10.1016/j.gloenvcha.2015.01.004)). All climate inputs come from [WorldClim 2.1](https://worldclim.org/), a widely used source of high-resolution, interpolated climate datasets based on weather station observations worldwide ([Fick & Hijmans, 2017](https://doi.org/10.1002/joc.5086)).
 
@@ -129,7 +129,7 @@ This series includes 12 monthly data points from [downscaled](https://worldclim.
 
 ### Setup
 
-To get started, ensure you have [NetLogo](https://ccl.northwestern.edu/netlogo) installed. This model was developed using NetLogo 7, so it is recommended to use this version or later.
+To get started, ensure you have [NetLogo](https://www.netlogo.org) installed. This model was developed using NetLogo 7.0, so it is recommended to use this version or later.
 
 The model relies on the GIS ([`gis`](https://github.com/NetLogo/GIS-Extension)), Pathdir ([`pathdir`](https://github.com/cstaelin/Pathdir-Extension)), String ([`string`](https://github.com/NetLogo/String-Extension)), and Time ([`time`](https://github.com/NetLogo/Time-Extension/)) NetLogo extensions. These are automatically installed when the model is run for the first time.
 
@@ -157,7 +157,7 @@ Refer to the `Info` tab in the model for additional details.
 
 ### Integrating with Other Models
 
-`LogoClim` can be integrated with other models using NetLogo's LevelSpace ([`ls`](https://ccl.northwestern.edu/netlogo/docs/ls.html)) extension. This extension enables parallel execution and data exchange between models. For an example of integrating `LogoClim` with another model, see the [FoodClim](https://github.com/sustentarea/foodclim) project.
+`LogoClim` can be integrated with other models using NetLogo's LevelSpace ([`ls`](https://www.netlogo.orgdocs/ls.html)) extension. This extension enables parallel execution and data exchange between models. For an example of integrating `LogoClim` with another model, see the [FoodClim](https://github.com/sustentarea/foodclim) project.
 
 ## How to Cite
 
@@ -216,7 +216,7 @@ You can also support the development of `LogoClim` by becoming a sponsor. Click 
 
 cat(
   badger::badge_license(
-    license = "GPL-3",
+    license = "GPLv3",
     color = "bd0000",
     url = "https://www.gnu.org/licenses/gpl-3.0"
   )
@@ -226,7 +226,7 @@ cat(
 ```text
 Copyright (C) 2025 Daniel Vartanian
 
-LogoClim is free software: you can redistribute it and/or modify it under the
+LogoClim is a free software: you can redistribute it and/or modify it under the
 terms of the GNU General Public License as published by the Free Software
 Foundation, either version 3 of the License, or (at your option) any later
 version.
