### 2025-07-30 2a910a2
Add patch changes to the documentation
