+++ README.qmd
@@ -157,7 +157,7 @@ Refer to the `Info` tab in the model for additional details.
 
 ### Integrating with Other Models
 
-`LogoClim` can be integrated with other models using NetLogo's LevelSpace ([`ls`](https://www.netlogo.orgdocs/ls.html)) extension. This extension enables parallel execution and data exchange between models. For an example of integrating `LogoClim` with another model, see the [FoodClim](https://github.com/sustentarea/foodclim) project.
+`LogoClim` can be integrated with other models using NetLogo's LevelSpace ([`ls`](https://ccl.northwestern.edu/netlogo/docs/ls.html)) extension. This extension enables parallel execution and data exchange between models. For an example of integrating `LogoClim` with another model, see the [FoodClim](https://github.com/sustentarea/foodclim) project.
 
 ## How to Cite
 
@@ -226,7 +226,7 @@ cat(
 ```text
 Copyright (C) 2025 Daniel Vartanian
 
-LogoClim is a free software: you can redistribute it and/or modify it under the
+LogoClim is free software: you can redistribute it and/or modify it under the
 terms of the GNU General Public License as published by the Free Software
 Foundation, either version 3 of the License, or (at your option) any later
 version.
