### 2025-08-04 f0c78fd
Add Brazil box dimensions
