### 2025-08-06 d29c78d
Update CRediT taxonomy roles in JOSS paper
