+++ .vscode/ltex.hiddenFalsePositives.en-US.txt
@@ -29,3 +29,4 @@
 {"rule":"EN_UNPAIRED_BRACKETS","sentence":"^\\Qhtmltools::tags$table(\n  htmltools::tags$tr(\n    htmltools::tags$td(\n      htmltools::tags$br(),\n      htmltools::tags$p(\n        htmltools::tags$a(\n          htmltools::tags$img(\n            src = \"images/cnpq-logo.svg\",\n            width = 150\n          ),\n          href = \"https://www.gov.br/cnpq/\"\n        ),\n        align = \"center\"\n      ),\n      htmltools::tags$br(),\n      width = \"30%\"\n    ),\n    htmltools::tags$td(\n      htmltools::tags$p(\n        \"This work was supported by the Department of Science and \",\n        \"Technology of the Secretariat of Science, Technology, and Innovation \",\n        \"and of the Health Economic-Industrial Complex (\",\n        htmltools::tags$a(\n          \"SECTICS\",\n          href = \"https://www.gov.br/saude/pt-br/composicao/sectics/\",\n          .noWS = \"outside\"\n        ),\n        \")  of the \",\n        htmltools::tags$a(\n          \"Ministry of Health\",\n          href = \"https://www.gov.br/saude/pt-br/composicao/sectics/\",\n          .noWS = \"outside\"\n        ), \" \",\n        \"of Brazil, and the National Council for Scientific and \",\n        \"Technological Development (\",\n        htmltools::tags$a(\n          \"CNPq\",\n          href = \"https://www.gov.br/cnpq/\",\n          .noWS = \"outside\"\n        ),\n        \") (grant no. 444588/2023-0)\"\n      ),\n      width = \"70%\"\n    )\n  )\n) |>\n  knitr::raw_html()\n```\\E$"}
 {"rule":"WHITESPACE_RULE","sentence":"^\\Qhtmltools::tags$table(\n  htmltools::tags$tr(\n    htmltools::tags$td(\n      htmltools::tags$br(),\n      htmltools::tags$p(\n        htmltools::tags$a(\n          htmltools::tags$img(\n            src = \"images/cnpq-logo.svg\",\n            width = 150\n          ),\n          href = \"https://www.gov.br/cnpq/\"\n        ),\n        align = \"center\"\n      ),\n      htmltools::tags$br(),\n      width = \"30%\"\n    ),\n    htmltools::tags$td(\n      htmltools::tags$p(\n        \"This work was supported by the Department of Science and \",\n        \"Technology of the Secretariat of Science, Technology, and Innovation \",\n        \"and of the Health Economic-Industrial Complex (\",\n        htmltools::tags$a(\n          \"SECTICS\",\n          href = \"https://www.gov.br/saude/pt-br/composicao/sectics/\",\n          .noWS = \"outside\"\n        ),\n        \")  of the \",\n        htmltools::tags$a(\n          \"Ministry of Health\",\n          href = \"https://www.gov.br/saude/pt-br/composicao/sectics/\",\n          .noWS = \"outside\"\n        ), \" \",\n        \"of Brazil, and the National Council for Scientific and \",\n        \"Technological Development (\",\n        htmltools::tags$a(\n          \"CNPq\",\n          href = \"https://www.gov.br/cnpq/\",\n          .noWS = \"outside\"\n        ),\n        \") (grant no. 444588/2023-0)\"\n      ),\n      width = \"70%\"\n    )\n  )\n) |>\n  knitr::raw_html()\n```\\E$"}
[…]
