+++ .vscode/ltex.disabledRules.en-US.txt
@@ -1,3 +1,5 @@
 PASSIVE_VOICE
 EN_QUOTES
 THREE_NN
+COMMA_PARENTHESIS_WHITESPACE
+DASH_RULE
