### 2025-08-19 #8553 @platipodium
> I am still involved with two other reviews for JOSS, so I fear I won't be able to dedicate too much time currently. But it does seem interesting!

Just finished that other review for most parts, so I'd be free and happy to review this @cheginit
