### 2025-08-21 #8845 @editorialbot
[REVIEW]: LogoClim: WorldClim in NetLogo
**Submitting author:** @danielvartan (<a href="http://orcid.org/0000-0001-7782-759X">Daniel Vartanian</a>)
**Repository:** https://github.com/sustentarea/logoclim
**Branch with paper.md** (empty if default branch):
**Version:** v2.2.1
