### 2025-08-21 ffa55b1
Remove `#wait` parameter from `go` procedure
