### 2025-08-21 3cc0174
Update Quarto notebooks
