+++ qmd/data-upload.qmd
@@ -17,7 +17,7 @@ This document is intended for use by the LogoClim project team. For related work
 
 This document provides a step-by-step guide for uploading WorldClim 2.1 data to the [LogoClim OSF repository](https://osf.io/2bpjh/). For information on obtaining the data, see the [WorldClim website](https://www.worldclim.org/).
 
-## Setting the Environment
+## Set the Environment
 
 ```{r}
 library(beepr)
@@ -27,6 +27,7 @@ library(fs)
 library(groomr) # github.com/danielvartan/groomr
 library(here)
 library(osfr)
+library(prettycheck) # github.com/danielvartan/prettycheck
 library(purrr)
 library(readr)
 library(rutils) # github.com/danielvartan/rutils
@@ -35,11 +36,11 @@ library(zip) # OS independent
 ```
 
 ```{r}
-source(here::here("R", "wc_license.R"))
-source(here::here("R", "wc_readme.R"))
+source(here("R", "wc_license.R"))
+source(here("R", "wc_readme.R"))
 ```
 
-## Setting the Initial Variables
+## Set the Initial Variables
 
 ### `params`
 
@@ -116,47 +117,47 @@ resolution <- "all"
 Check the paths to the directories where the data will be stored.
 
 ```{r}
-raw_data_dir <- here::here("data-raw")
+raw_data_dir <- here("data-raw")
 ```
 
 ```{r}
-wc_raw_data_dir <- fs::path(raw_data_dir, "worldclim")
+raw_data_wc_dir <- path(raw_data_dir, "worldclim")
 ```
 
 ```{r}
-wc_raw_data_series_dir <- fs::path(wc_raw_data_dir, series)
+raw_data_wc_series_dir <- path(raw_data_wc_dir, series)
 ```
 
 ```{r}
-wc_raw_data_series_res_dir <- fs::path(
-  wc_raw_data_dir,
+raw_data_wc_series_res_dir <- path(
+  raw_data_wc_dir,
   series,
-  resolution |> stringr::str_replace_all("\\.", "\\-")
+  resolution |> str_replace_all("\\.", "\\-")
 )
 ```
 
 ```{r}
-data_dir <- here::here("data")
+data_dir <- here("data")
 ```
 
 ```{r}
-data_series_dir <- fs::path(data_dir, series)
+data_series_dir <- path(data_dir, series)
 ```
 
 ```{r}
 dirs <- c(
-  raw_data_dir, wc_raw_data_dir, wc_raw_data_series_dir,
-  wc_raw_data_series_res_dir, data_dir, data_series_dir
+  raw_data_dir, raw_data_wc_dir, raw_data_wc_series_dir,
+  raw_data_wc_series_res_dir, data_dir, data_series_dir
 )
 
 for (i in dirs) {
-  if (!dir.exists(i)) {
-    fs::dir_create(i, recurse = TRUE)
+  if (!dir_exists(i)) {
+    dir_create(i, recurse = TRUE)
   }
 }
 ```
 
-## Zipping the Files
+## Zip the Files
 
 ```{r}
 #| eval: false
@@ -192,11 +193,11 @@ if (length(series) == 1) {
 
 zip_file <-
   data_dir |>
-  fs::path(
+  path(
       paste0(
       country_code,
       ifelse(!is.null(country_suffix), paste0("-", country_suffix), ""),
-      paste0("-", stringr::str_replace_all(resolution, "\\.", "\\-")),
+      paste0("-", str_replace_all(resolution, "\\.", "\\-")),
[…]
