+++ qmd/data-munging.qmd
@@ -16,7 +16,7 @@ This document provides a step-by-step guide to transform the WorldClim 2.1 data 
 This process may take some time to complete. Please be patient.
 :::
 
-## Setting the Environment
+## Set the Environment
 
 ::: {.callout-important}
 Notice date: 2025-04-09
@@ -43,11 +43,11 @@ library(zip) # OS independent
 ```
 
 ```{r}
-source(here::here("R", "wc_license.R"))
-source(here::here("R", "logoclim_readme.R"))
+source(here("R", "wc_license.R"))
+source(here("R", "logoclim_readme.R"))
 ```
 
-## Setting the Initial Variables
+## Set the Initial Variables
 
 ### `params`
 
@@ -173,42 +173,42 @@ data
 The following settings assumes that the data will be stored in the `data` directory of the project, but you can change it to any other directory. If you do, make sure to change the model `data-path` variable accordingly.
 
 ```{r}
-raw_data_dir <- here::here("data-raw")
+raw_data_dir <- here("data-raw")
 ```
 
 ```{r}
-wc_raw_data_dir <- fs::path(raw_data_dir, "worldclim")
+raw_data_wc_dir <- path(raw_data_dir, "worldclim")
 ```
 
 ```{r}
-wc_raw_data_series_dir <- fs::path(wc_raw_data_dir, series)
+raw_data_wc_series_dir <- path(raw_data_wc_dir, series)
 ```
 
 ```{r}
-wc_raw_data_series_res_dir <- fs::path(
-  wc_raw_data_dir,
+raw_data_wc_series_res_dir <- path(
+  raw_data_wc_dir,
   series,
-  resolution |> stringr::str_replace_all("\\.", "\\-")
+  resolution |> str_replace_all("\\.", "\\-")
 )
 ```
 
 ```{r}
-data_dir <- here::here("data")
+data_dir <- here("data")
 ```
 
 ```{r}
-data_series_dir <- fs::path(data_dir, series)
+data_series_dir <- path(data_dir, series)
 ```
 
 ```{r}
 dirs <- c(
-  raw_data_dir, wc_raw_data_dir, wc_raw_data_series_dir,
-  wc_raw_data_series_res_dir, data_dir, data_series_dir
+  raw_data_dir, raw_data_wc_dir, raw_data_wc_series_dir,
+  raw_data_wc_series_res_dir, data_dir, data_series_dir
 )
 
 for (i in dirs) {
-  if (!dir.exists(i)) {
-    fs::dir_create(i, recurse = TRUE)
+  if (!dir_exists(i)) {
+    dir_create(i, recurse = TRUE)
   }
 }
 ```
@@ -236,7 +236,7 @@ Change to the desired country code (ISO 3166-1 alpha-3).
 ```{r}
 #| eval: false
 
-shape <- geodata::gadm(
+shape <- gadm(
   country = country_code,
   level = 0,
   path = raw_data_dir
@@ -251,12 +251,12 @@ shape <- geodata::gadm(
 #| eval: false
 
 shape <-
-  geodata::gadm(
+  gadm(
     country = "usa",
     level = 1,
     path = raw_data_dir
   ) %>%
-  terra::subset(., !.$NAME_1 %in% c("Alaska", "Hawaii"))
+  subset(., !.$NAME_1 %in% c("Alaska", "Hawaii"))
 ```
 
 ##### Portugal
@@ -265,12 +265,12 @@ shape <-
 #| eval: false
 
 shape <-
-  geodata::gadm(
+  gadm(
     country = "prt",
     level = 1,
     path = raw_data_dir
   ) %>%
[…]

[…]
