### 2025-08-21 sustentarea/logoclim#8 @platipodium
Consider streamlining supplementary sections in JOSS software paper
Related https://github.com/openjournals/joss-reviews/issues/8845

JOSS has a word limit of 1000 and your paper is on the higher end of that.  In contrast to many other papers, your supplementary sections including Acknowledgements, Author Contribution, Additional Information and References are very long, constituting 50% of the text.

While I see the necessity and good practices in having these, they seem to be too much for a JOSS paper, but the editor might be of a different opinion.

Please clarify.
