### 2025-08-21 #8845 @platipodium
It seems the `clock` tool is - again - lagging behind.  I had added the netlogo part, but it does not yet recognize nlogox format, which it mistakes for generic xml.

```
Language                     files          blank        comment           code
XML                              1             89              0           1280
NetLogo                         60            261            183           1271
```

So really your source code is more or less 2400 LOC
