### 2025-08-22 sustentarea/logoclim#8 @cheginit
I agree with @platipodium suggestions. Since you have already referenced the CMIP6 and WorldClim, you've already credited their work. So the first three paragraphs are not really needed. Per the suggestion of @platipodium, move the disclaimer to the acknowledgement section, and if you like, you can modify the disclaimer to thank them in few words for making the dataset open, then continue with the rest. We do not require a contributions section, please remove this section too.
