### 2025-08-22 2af7505
Streamline supplementary sections of the JOSS paper #8
