### 2025-08-22 9d86da4
Rephrase the "Acknowledgements" section #8
