### 2025-08-29 sustentarea/logoclim#9 @jamesdamillington
JOSS paper: verifying functionality
Related to https://github.com/openjournals/joss-reviews/issues/8845

One of the review questions is:

> "Are there automated tests or manual steps described so that the functionality of the software can be verified?"

Currently I feel I can't answer in the affirmative for this. While the documentation appropriately explains how to Setup and Run the model, it doesn't provide any guidance on how the user will know if the model is functioning properly. Automated tests may not be possible, but could you suggest some steps, and their outcomes, that could be verified by the user? e.g. for a given downloaded dataset/variable what should expect to see after clicking Setup then Go?
