### 2025-08-29 sustentarea/logoclim#10 @jamesdamillington
JOSS Paper: functionality and example usage
Related to https://github.com/openjournals/joss-reviews/issues/8845 and https://github.com/sustentarea/logoclim/issues/9

I feel I can't answer in the affirmative to two related paper review questions:

> - Have the functional claims of the software been confirmed?
> - Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

It seems to me that the main functional claim of the software is that **this model can be integrated with _other_ NetLogo models to facilitate use and access to climate data**. That's a fair aim, but there are no examples here of how to do such integration.

I see from the _Motivation_ section of the paper that this project arose from another, [FoodClim](https://github.com/sustentarea/foodclim). I tried downloading and running that model to see how I could integrate LogoClim, but I received an error after what I thought was appropriate setup (incl. setting the directory of LogoClim). See detail [here](https://github.com/sustentarea/foodclim/issues/1). So, currently I cannot verify the functional claim of the software (to enable integration of climate data into _other_ models)

Rather than using the FoodClim model to achieve the two points above, I wonder if the authors might provide a second simple 'test model' in _this_ repo to demonstrate how integration works. It does not need to be a complicated model, maybe a simple 'species distribution model' that uses a pre-specified equation to use two or three of the climate variables to predict presence for a given area (it doesn't have to be a real model, just something that demonstrates the functionality)? This 'test model' might also help to [this issue](https://github.com/sustentarea/logoclim/issues/9)
