### 2025-08-29 sustentarea/logoclim#11 @jamesdamillington
JOSS paper: state of the field
Related to https://github.com/openjournals/joss-reviews/issues/8845

One of the review questions is:

> Do the authors describe how this software compares to other commonly-used packages?

I can't currently see this in the paper. I'm not sure there is anything similar for NetLogo (have you checked?), but there are certainly similar packages for other platforms/languages. For example, for R there is [_climenv_](https://doi.org/10.1111/jvs.13215). So, I think a brief sentence to highlight similar packages for other platforms/languages is needed (and if there's nothing similar for NetLogo you could briefly state something like, "_we know of no similar models for NetLogo_").
