### 2025-08-29 #8845 @jamesdamillington
@danielvartan This looks like it could be a useful tool for others, and seems to be well constructed. Well done! However, there are a few issues before I can complete my checklist. See the issues I have raised in the repo:
- https://github.com/sustentarea/logoclim/issues/9
- https://github.com/sustentarea/logoclim/issues/10
- https://github.com/sustentarea/logoclim/issues/11
