+++ README.qmd
@@ -226,7 +226,7 @@ cat(
 ```text
 Copyright (C) 2025 Daniel Vartanian
 
-LogoClim is free software: you can redistribute it and/or modify it under the
+{LogoClim} is free software: you can redistribute it and/or modify it under the
 terms of the GNU General Public License as published by the Free Software
 Foundation, either version 3 of the License, or (at your option) any later
 version.
