+++ SECURITY.md
@@ -1,3 +1,3 @@
 # Security Policy
 
-To report a security issue, please email hi@danielvartan.com with a description of the issue, the steps you took to create the issue, affected versions, and, if known, mitigations for the issue. This project follows a 90 day disclosure timeline.
+To report a security issue, please email `hi@danielvartan.com` with a description of the issue, the steps you took to create the issue, affected versions, and, if known, mitigations for the issue. This project follows a 90 day disclosure timeline.
