+++ README.md
@@ -230,7 +230,7 @@ a donation. Please mention `LogoClim` in your donation message.
 ``` text
 Copyright (C) 2025 Daniel Vartanian
 
-LogoClim is free software: you can redistribute it and/or modify it under the
+{LogoClim} is free software: you can redistribute it and/or modify it under the
 terms of the GNU General Public License as published by the Free Software
 Foundation, either version 3 of the License, or (at your option) any later
 version.
