### 2025-09-13 sustentarea/logoclim#9 @danielvartan
Hi @jamesdamillington,

I'm developing a Quarto notebook that uses the R package [nlrx](https://docs.ropensci.org/nlrx/) (from [rOpenSci](https://ropensci.org/)) to programmatically run NetLogo experiments from R, and then uses the R package [testthat](https://testthat.r-lib.org/) to define expectations and validate results. It’s not a full testing suite, but it can serve a similar purpose.

I plan to work on this over the next few days. I just launched the `Logônia` model to address issue #10.
