### 2025-09-13 sustentarea/logoclim#11 @danielvartan
Hi @jamesdamillington,

At the moment, there’s nothing similar available for NetLogo. But you’re right—it’s worth highlighting other packages with comparable functionality. I’ll work on this over the next few weeks.
