### 2025-09-13 #8845 @danielvartan
Hi everyone,

Thank you @platipodium for the suggestion. I'm working on a similar approach (see https://github.com/sustentarea/logoclim/issues/9#issuecomment-3287797147).

Here are a few updates:

Uri Wilensky [recently announced](https://forum.netlogo.org/t/netlogo-7-released/636) the launch of the stable version of NetLogo 7. Since I had already adapted `LogoClim` to the beta release (https://github.com/openjournals/joss-reviews/issues/8553#issuecomment-3134039761), the model runs smoothly on the stable version as well.

Following a suggestion from @jamesdamillington regarding https://github.com/sustentarea/logoclim/issues/10, I created a test model that demonstrates how to integrate `LogoClim` through the LevelSpace NetLogo extension. I called it `Logônia`, a model that simulates the growth response of a fictional plant under different climatic conditions. The model is already available on [GitHub](https://github.com/sustentarea/logonia) and the [CoMSES Network](https://www.comses.net/codebases/4f2be13a-3957-4537-bf64-3fad96ba271f/). Please let me know if this fully addresses the issue mentioned.

![Image](https://github.com/user-attachments/assets/7bde927f-2d8d-471c-b671-a776e4f6d0ff)

![Image](https://github.com/user-attachments/assets/98b8735b-17b5-4990-9d29-2fd5faffb6ea)

![Image](https://github.com/user-attachments/assets/3c116153-fc4c-464d-8d58-32b5847eed67)

I’m still working on https://github.com/sustentarea/logoclim/issues/9 and https://github.com/sustentarea/logoclim/issues/11, and I plan to resolve them by the end of the month.
