### 2025-09-13 sustentarea/logoclim#10 @danielvartan
Hi @jamesdamillington,

As you suggested, I created a test model to address this issue. I named it `Logônia`, a model that simulates the growth response of a fictional plant under varying climatic conditions. The model is already available on [GitHub](https://github.com/sustentarea/logonia) and the [CoMSES Network](https://www.comses.net/codebases/4f2be13a-3957-4537-bf64-3fad96ba271f/). Please let me know if this fully addresses the two points you mentioned.

![Image](https://github.com/user-attachments/assets/1707a923-3c90-48df-a089-3217c29b08bc)

![Image](https://github.com/user-attachments/assets/557773d7-5587-4e0c-b92f-e1a2bea97dc0)

![Image](https://github.com/user-attachments/assets/b57f1c2d-9a5a-44a8-8e68-8222a0909c24)
