### 2025-09-13 f6c4963
Add GitHub Action for automatic publishing
