### 2025-09-13 1e9c32f
Add `Logônia` to the documentation
