+++ NEWS.md
@@ -1,17 +1,19 @@
 # LogoClim 2.0.0.9000 (development version) (NetLogo 7.0.0)
 
-- Removed `#wait` parameters from `go` procedure, as it is no longer necessary
-- Updated Quarto notebooks
+- Removed `#wait` parameters from `go` procedure, as it is no longer necessary.
+- Updated Quarto notebooks.
+- Greatly improved the contrast of patch labels in the `show-value` procedure.
+- Added the [`Logônia`](https://github.com/sustentarea/logonia) model as reference for `LogoClim` integration.
 
 # LogoClim 2.0.0 (NetLogo 7.0.0)
 
-- Updated the model for compatibility with NetLogo 7.0.0
-- Resized interface widgets to match the new NetLogo standard dimensions
-- Changed the `LogoClim` license from MIT to GPLv3
-- Removed the `transition-seconds` slider, as it is no longer necessary
-- Removed the `adjust-world-size?` slider; world size adjustment is now always enabled by default
-- Introduced the global variable `plot-max-y-range` to optimize computations
-- Revised documentation to reflect updates
+- Updated the model for compatibility with NetLogo 7.0.0.
+- Resized interface widgets to match the new NetLogo standard dimensions.
+- Changed the `LogoClim` license from MIT to GPLv3.
+- Removed the `transition-seconds` slider, as it is no longer necessary.
+- Removed the `adjust-world-size?` slider; world size adjustment is now always enabled by default.
+- Introduced the global variable `plot-max-y-range` to optimize computations.
+- Revised documentation to reflect updates.
 
 # LogoClim 1.0.0 (NetLogo 6.4.0)
 
@@ -24,15 +26,15 @@ First stable release. 🎉
 
 # LogoClim 0.0.0.9013 (Pre-Release) (NetLogo 6.4.0)
 
-- Improved the documentation
-- Removed all dependencies on the `R` programming language and its packages
-- Improved `setup-world` to address bleeding issues
-- Persistent world bleeding is now converted to `NaN` values
-- Enhanced Quarto notebooks to fix dateline issues
+- Improved the documentation.
+- Removed all dependencies on the `R` programming language and its packages.
+- Improved `setup-world` to address bleeding issues.
+- Persistent world bleeding is now converted to `NaN` values.
+- Enhanced Quarto notebooks to fix dateline issues.
 - Automated the generation of README and LICENSE files in the
-  Quarto notebooks
+  Quarto notebooks.
 - Removed `patch-px-size` slider and added `adjust-world-size?` slider for
-  automatic world size adjustment
+  automatic world size adjustment.
 - Removed automatic adjustment of `start-year`. An error is now raised if
   `start-year` is not set to a valid value.
 - Removed unnecessary dependencies and refactored code structure for improved
@@ -41,15 +43,15 @@ First stable release. 🎉
 # LogoClim 0.0.0.9010 (Pre-Release) (NetLogo 6.4.0)
 
 - Fixed an issue with Windows file paths for improved cross-platform
-compatibility
+compatibility.
 
 # LogoClim 0.0.0.9009 (Pre-Release) (NetLogo 6.4.0)
 
-- Enhanced Quarto notebooks for improved readability and clarity
-- Refactored codebase to increase modularity and maintainability
-- Resolved world bleeding issues at higher resolutions
-- Updated charts to display whole numbers
-- Expanded documentation
+- Enhanced Quarto notebooks for improved readability and clarity.
+- Refactored codebase to increase modularity and maintainability.
+- Resolved world bleeding issues at higher resolutions.
[…]
