+++ CODE_OF_CONDUCT.md
@@ -60,7 +60,7 @@ representative at an online or offline event.
 
 Instances of abusive, harassing, or otherwise unacceptable behavior may be
 reported to the community leaders responsible for enforcement at
-[hi@danielvartan.com](mailto:hi@danielvartan.com).
+[daniel.azevedo@alumni.usp.br](mailto:daniel.azevedo@alumni.usp.br).
 All complaints will be reviewed and investigated promptly and fairly.
 
 All community leaders are obligated to respect the privacy and security of the
