### 2025-09-14 0342d9c
Resolves #12 and add other patches
