+++ .vscode/ltex.disabledRules.en-US.txt
@@ -1,5 +1,0 @@
-PASSIVE_VOICE
-EN_QUOTES
-THREE_NN
-COMMA_PARENTHESIS_WHITESPACE
-DASH_RULE
