+++ .vscode/ltex.hiddenFalsePositives.en-US.txt
@@ -1,32 +1,0 @@
-{"rule":"COMMA_PARENTHESIS_WHITESPACE","sentence":"^\\Qcrop_shape <-\n  geodata::gadm(\n    country = \"usa\",\n    level = 1,\n    path = raw_data_dir\n  ) %>%\n  terra::subset(., !.$NAME_1 %in% c(\"Alaska\", \"Hawaii\"))\n```\\E$"}
-{"rule":"PASSIVE_VOICE","sentence":"^\\QImproved \\E(?:Dummy|Ina|Jimmy-)[0-9]+\\Q to utilize the first three files in the series and address bleeding issues\nPersistent world bleeding is now converted to \\E(?:Dummy|Ina|Jimmy-)[0-9]+\\Q values\nEnhanced Quarto notebooks to fix dateline issues\nAutomated the generation of README and LICENSE files in the Quarto notebooks\\E$"}
-{"rule":"MORFOLOGIK_RULE_EN_US","sentence":"^\\Q---\nparams:\n  series: \"historical-climate-data\"\n  resolution: \"10m\"\n---\\E$"}
-{"rule":"THREE_NN","sentence":"^\\QFixed an issue with year intervals when using the Future climate data series.\\E$"}
-{"rule":"EN_QUOTES","sentence":"^\\Q```{r}\n#| label: setup\n#| include: false\\E$"}
-{"rule":"COMMA_PARENTHESIS_WHITESPACE","sentence":"^\\Qcat(\n  badger::badge_repostatus(\"active\"),\n  badger::badge_custom(\n    x = \"CoMSES Network\",\n    y = \"not published\",\n    color = \"F5514D\", # \"1284C5\"\n    url = \"https://www.comses.net/\"\n  ),\n  badger::badge_custom(\n    x = \"OSF DOI\",\n    y = \"10.17605/OSF.IO/EAPZU\",\n    color = \"1284C5\",\n    url = \"https://doi.org/10.17605/OSF.IO/EAPZU\"\n  ),\n  paste0(\n    \"[![\",\n    \"fair-software.eu\",\n    \"](\",\n    \"https://img.shields.io/badge/\",\n    \"fair--software.eu\", \"-\",\n    \"%E2%97%8F%20%20%E2%97%8F%20%20%E2%97%8B%20%20%E2%97%8F%20%20%E2%97%8B\",\n    \"-\",\n    \"orange\",\n    \")](\",\n    \"https://fair-software.eu\",\n    \")\"\n  ),\n  paste0(\n    \"[![FAIR checklist badge](https://fairsoftwarechecklist.net/badge.svg)]\",\n    \"(https://fairsoftwarechecklist.net/v0.2?f=21&a=32113&i=32001&r=123)\"\n  ),\n  badger::badge_license(\n    license = \"MIT\",\n    color = \"green\",\n    url = \"https://choosealicense.com/licenses/mit/\"\n  ),\n  badger::badge_custom(\n    x = \"Contributor%20Covenant\",\n    y = \"2.1\",\n    color = \"4baaaa\",\n    url = \"CODE_OF_CONDUCT.md\"\n  )\n)\n```\n<!-- badges: end -->\\E$"}
[…]
