+++ .vscode/ltex.dictionary.en-US.txt
@@ -1,44 +1,0 @@
-Clim
-LogoClim
-WorldClim
-GCMs
-beepr
-groomr
-lubridate
-readr
-stringr
-asis
-repostatus
-CoMSES
-url
-NetLogo
-SSPs
-Fick
-Hijmans
-&nbsp
-bioclimatic
-downscaled
-gis
-Pathdir
-pathdir
-logoclim
-nlogo
-nlogox
-LevelSpace
-FoodClim
-gfm
-Vartanian
-@Misc
-doi
-ESGF
-knitr
-htmltools
-CRediT
-CNPq
-SECTICS
-Sustentarea
-Intercomparison
-Programme
-submodel
-newpage
-autoref
