### 2025-09-14 79adb1a
Change string parameters and interface text to title case
