### 2025-09-14 ef87223
Fix local variable names
