+++ NEWS.md
@@ -1,4 +1,4 @@
-# LogoClim 2.0.0.9000 (development version) (NetLogo 7.0.0)
+## v2.0.0.9000 (Development Version)
 
 - Updated Quarto notebooks.
 - Removed `#wait` parameters from `go` procedure, as it is no longer necessary.
@@ -11,9 +11,10 @@
 - Changed string parameters and interface text to title case.
 - Revised documentation to reflect updates.
 
-# LogoClim 2.0.0 (NetLogo 7.0.0)
+## v2.0.0
 
-- Updated the model for compatibility with NetLogo 7.0.0.
+- **Breaking change**: Updated the model for compatibility with **NetLogo 7.0.0**.
+
 - Resized interface widgets to match the new NetLogo standard dimensions.
 - Changed the `LogoClim` license from MIT to GPLv3.
 - Removed the `transition-seconds` slider, as it is no longer necessary.
@@ -21,16 +22,16 @@
 - Introduced the global variable `plot-max-y-range` to optimize computations.
 - Revised documentation to reflect updates.
 
-# LogoClim 1.0.0 (NetLogo 6.4.0)
+## v1.0.0
 
 First stable release. 🎉
 
-# LogoClim 0.0.0.9015 (Pre-Release) (NetLogo 6.4.0)
+## v0.0.0.9015 (Pre-Release)
 
 - Added variations (e.g., ACCESS-ESM1-5) and additional (e.g., CanESM5) Global Climate Models (GCMs) as selectable options in `global-climate-model`. WorldClim provides a dedicated webpage for this data, available [here](https://www.worldclim.org/data/cmip6_all/cmip6_clim2.5m.html).
 - Fixed an issue with year intervals when using the *Future Climate Data* series
 
-# LogoClim 0.0.0.9013 (Pre-Release) (NetLogo 6.4.0)
+## v0.0.0.9013 (Pre-Release)
 
 - Improved the documentation.
 - Removed all dependencies on the `R` programming language and its packages.
@@ -46,23 +47,27 @@ First stable release. 🎉
 - Removed unnecessary dependencies and refactored code structure for improved
   maintainability.
 
-# LogoClim 0.0.0.9010 (Pre-Release) (NetLogo 6.4.0)
+## v0.0.0.9010 (Pre-Release)
 
 - Fixed an issue with Windows file paths for improved cross-platform
 compatibility.
 
-# LogoClim 0.0.0.9009 (Pre-Release) (NetLogo 6.4.0)
+## v0.0.0.9009 (Pre-Release)
 
 - Enhanced Quarto notebooks for improved readability and clarity.
 - Refactored codebase to increase modularity and maintainability.
-- Resolved world bleeding issues at higher resolutions.
+- Fixed world bleeding issues at higher resolutions.
 - Updated charts to display whole numbers.
 - Expanded documentation.
 
-# LogoClim 0.0.0.9004 (Pre-Release) (NetLogo 6.4.0)
+## v 0.0.0.9064 (Pre-Release)
+
+- Improved internal mechanisms for better performance and reliability.
+
+## v 0.0.0.9004 (Pre-Release)
 
 First pre-release. 🎉
 
-# LogoClim 0.0.0.9000 (NetLogo 6.4.0)
+## v0.0.0.9000
 
 - Added a `NEWS.md` file to track changes to the model.
