+++ images/logoclim-interface.md
@@ -54,8 +54,7 @@ wmctrl -r NetLogo -e 0,112,75,1696,807
 - France: HCD, 30s, 1970-2000, Water vapor pressure
 
 ```netlogo
-go true true
-# Clear the command center after
+go
 ```
 
 ### Render
