+++ README.md
@@ -24,7 +24,7 @@ GPLv3](https://img.shields.io/badge/license-GPLv3-bd0000.svg)](https://www.gnu.o
 and visualizing global climate conditions. It allows researchers to
 integrate high-resolution climate data into agent-based models,
 supporting reproducible research in ecology, agriculture, environmental
-science, and other fields that rely on climate data integration.
+sciences, and other fields that rely on climate data.
 
 The model utilizes raster data to represent climate variables such as
 temperature and precipitation over time. It incorporates historical data
@@ -45,10 +45,10 @@ example of how to integrate `LogoClim` into your model.
 > [![GitHub repo
 > stars](https://img.shields.io/github/stars/sustentarea/logoclim)](https://github.com/sustentarea/logoclim/)
 
+![LogoClim Interface](images/logoclim-interface.gif)
+
 > [!IMPORTANT]
 > `LogoClim` is an independent project with no affiliation to [WorldClim](https://worldclim.org/) or its developers. Users should be aware that WorldClim datasets are freely available for academic and other non-commercial use only. Any use of WorldClim data within `LogoClim` must comply with [WorldClim's licensing terms](https://worldclim.org/about.html).
-
-![LogoClim Interface](images/logoclim-interface.gif)
 
 ## How It Works
 
@@ -59,35 +59,22 @@ colors based on the data values. The results can be visualized on a map,
 accompanied by plots that display the mean, minimum, maximum, and
 standard deviation of the selected variable over time.
 
-### Color Scale
-
-The model uses a color scale ranging from black (representing the lowest
-value) to white (representing the highest value). Users can adjust the
-thresholds for these colors using the **`black-value`** and
-**`white-value`** sliders. Alternatively, users can set the black or
-white color to automatically represent the minimum or maximum value of
-the current data by toggling the **`black-min`** and **`white-max`**
-switches. By default, the black threshold is set to 0, and the white
-threshold corresponds to the maximum value of the current data.
-
 ### Data Series
 
-In addition to latitude and longitude data points, `LogoClim` supports
-simulation with all three climate data series provided by [WorldClim
-2.1](https://worldclim.org/): long-term historical climate averages,
-historical monthly weather, and future climate projections. Each series
-is available at multiple spatial resolutions (from 10 minutes (~340 km²
-at the equator) to 30 seconds (~1 km² at the equator)) and can be
-selected within the model interface to fit your research needs. More
-information about each series can be found in the WorldClim website.
+The model supports simulation with all three climate data series
+provided by [WorldClim 2.1](https://worldclim.org/). Each series is
+available at multiple spatial resolutions (from 10 minutes (~340 km² at
+the equator) to 30 seconds (~1 km² at the equator)) and can be selected
+within the model interface to fit your research needs. More information
+about each series can be found in the WorldClim website.
 
 #### Historical Climate Data
 
 This series includes only 12 monthly data points representing long-term
 average climate conditions for the period 1970-2000. It provides
 averages on minimum, mean, and maximum temperature, precipitation, solar
-radiation, wind speed, vapor pressure, elevation, and on bioclimatic
-variables.
+radiation, wind speed, vapor pressure, elevation, and on [bioclimatic
+variables](https://www.worldclim.org/data/bioclim.html).
 
 #### Historical Monthly Weather Data
 
@@ -112,7 +99,9 @@ projections cover four
 [SSPs](https://climatedata.ca/resource/understanding-shared-socio-economic-pathways-ssps/):
 126, 245, 370, and 585, with data available for average minimum
[…]
