### 2025-09-15 c585c8e
Update `README.md` with a new interface animation
