### 2025-09-16 3d41ab5
Convert `NaN` values to `false`
