### 2025-09-16 a569003
Update JOSS paper
