+++ paper/paper.md
@@ -1,22 +1,20 @@
 ---
-title: "LogoClim: WorldClim in NetLogo"
+title: "{LogoClim}: WorldClim in NetLogo"
 tags:
   - Agent-Based Modeling
   - Climate Change
   - Climate Data Visualization
-  - Climate Model Integration
   - Climate Projections
-  - Climate Simulation
+  - Climate Simulations
   - CMIP6
   - Complex Systems
   - Complexity Science
-  - Data Interoperability
   - Environmental Sciences
-  - Environmental Simulation
   - Future Climate Scenarios
   - Geospatial Analysis
   - Historical Climate Data
   - LevelSpace
+  - Logônia
   - Models
   - NetLogo
   - Parallel Execution
@@ -25,10 +23,9 @@ tags:
   - Shared Socioeconomic Pathways
   - Simulations
   - Spatial Analysis
-  - Spatial Resolution
   - SSPs
   - Time Series
-  - Worldclim
+  - WorldClim
 authors:
   - given-names: Daniel
     surname: Vartanian
@@ -51,17 +48,17 @@ affiliations:
   index: 2
 - name: School of Public Health, University of São Paulo, São Paulo, Brazil
   index: 3
-date: 22 August 2025
+date: 16 September 2025
 bibliography: paper.bib
 ---
 
 # Summary
 
-`LogoClim` is a [NetLogo](https://www.netlogo.org) model for simulating and visualizing global climate conditions. It allows researchers to integrate high-resolution climate data into agent-based models, supporting reproducible research in ecology, agriculture, environmental science, and other fields that rely on climate data integration.
+`LogoClim` is a [NetLogo](https://www.netlogo.org) model for simulating and visualizing global climate conditions. It allows researchers to integrate high-resolution climate data into agent-based models, supporting reproducible research in ecology, agriculture, environmental sciences, and other fields that rely on climate data.
 
-The model utilizes raster data to represent climate variables such as temperature and precipitation over time. It incorporates historical data (1951-2024) and future climate projections (2021-2100) derived from global climate models under various Shared Socioeconomic Pathways ([SSPs](https://climatedata.ca/resource/understanding-shared-socio-economic-pathways-ssps/)) [@oneill2017]. All climate inputs come from [WorldClim 2.1](https://worldclim.org/), a widely used source of high-resolution, interpolated climate datasets based on weather station observations worldwide [@fick2017a], available for academic and other non-commercial use.
+The model utilizes raster data to represent climate variables such as temperature and precipitation over time. It incorporates historical data (1951-2024) and future climate projections (2021-2100) derived from global climate models under various Shared Socioeconomic Pathways ([SSPs](https://climatedata.ca/resource/understanding-shared-socio-economic-pathways-ssps/)) [@oneill2017]. All climate inputs come from [WorldClim 2.1](https://worldclim.org/), a widely used source of high-resolution, interpolated climate datasets based on weather station observations worldwide [@fick2017], available for academic and other non-commercial use.
 
-`LogoClim` follows the FAIR Principles for Research Software [@barker2022] and is openly available on the [CoMSES Network](https://www.comses.net/codebases/bccd451f-76a4-408a-85fd-c5024359ba9a/) and [GitHub](https://github.com/sustentarea/logoclim). \autoref{fig:logoclim-interface-1} and \autoref{fig:logoclim-interface-2} illustrate the model's interface and functionality.
[…]
