### 2025-09-16 sustentarea/logoclim#11 @danielvartan
Hi @jamesdamillington,

I updated the [JOSS paper](https://github.com/sustentarea/logoclim/blob/main/paper/paper.pdf) with some minor changes, including the following addition at the end of the *Statement of need* section:

> While other programming languages, such as [R](https://www.r-project.org/), offer tools like the `geodata` package (Hijmans et al., [2024](https://CRAN.R-project.org/package=geodata)), there are currently no equivalent tools providing this functionality for NetLogo.

I thought it would be more appropriate to cite the [`geodata`](https://CRAN.R-project.org/package=geodata) package, since it was developed by one of the WorldClim authors.

Please let me know if this fully addresses the issue.
