### 2025-09-16 909e1e0
Update JOSS paper images
