### 2025-09-16 e81382c
Fix Hijmans reference
