### 2025-09-16 0f2cfee
Update CoMSES Network badge
