+++ README.qmd
@@ -33,7 +33,7 @@ cat(
   badger::badge_repostatus("active"),
   badger::badge_custom(
     x = "CoMSES Network",
-    y = "2.0.0",
+    y = "2.1.0",
     color = "1284C5",
     url = paste0(
       "https://www.comses.net/codebases/",
