+++ NEWS.md
@@ -1,4 +1,4 @@
-## v2.0.0.9000 (Development Version)
+## v2.1.0.9000 (Development Version)
 
 ## v2.1.0 (2025-09-16)
 
