### 2025-09-16 2e7ecdd
Add patches to R scripts
