### 2025-09-16 074c48d
Update hex logo
