### 2025-09-16 sustentarea/logoclim#11 @jamesdamillington
@danielvartan Yes, this looks good, I've checked this item off the review checklist and you can close this issue
