### 2025-09-16 59b2831
Fix inversion of latitude/longitude values
