### 2025-09-17 sustentarea/logoclim#8 @danielvartan
Hi @platipodium and @cheginit,

Are you satisfied with these changes? Can I close the issue?
