### 2025-09-18 sustentarea/logoclim#8 @platipodium
> Hi [@platipodium](https://github.com/platipodium) and [@cheginit](https://github.com/cheginit),
>
> Are you satisfied with these changes? Can I close the issue?

I am happy with these changes!  I would even go further and include the Motivation section (which is really only one paragraph, and rarely see in JOSS ) into the Statement of Need as its last paragraph.
