### 2025-09-18 #8845 @platipodium
I noticed your unusually title settings in curly braces `{LogoClim}` and `{Logônia}`.   This is distracting in the way that it is not clear which one is the name of your software: `{LogoClim}` or `LogoClim` ??  What do the others think @cheginit @jamesdamillington ?
