### 2025-09-18 sustentarea/logoclim#14 @platipodium
Add quarto instructions and R dependency management
Related https://github.com/openjournals/joss-reviews/issues/8845

I just tried to use the quarto notebooks for downloading data and was puzzeld

* for the lack of instructions what quarto is and how to use it.
* for the lack of dependency management with R.

Now I have used both Quarto and R in the past, but getting this to run is very tedious without instructions.
