### 2025-09-18 sustentarea/logoclim#17 @platipodium
Warn more clearly about space demand of WorldClim data
https://github.com/openjournals/joss-reviews/issues/8845

The sheer size of even the low-resolution data (1 GB for Europe at 10minutes) is overwhelming.  There should be more transparent information on the disk space demand for using logoclim with any of that data.
