### 2025-09-18 #8845 @danielvartan
> I noticed your unusually title settings in curly braces `{LogoClim}` and `{Logônia}`. This is distracting in the way that it is not clear which one is the name of your software: `{LogoClim}` or `LogoClim` ?? What do the others think [@cheginit](https://github.com/cheginit) [@jamesdamillington](https://github.com/jamesdamillington) ?

This is a tacit convention in the R community to distinguish package names in plain text (i.e., outside Markdown).

It also serves other purposes. For example, R package names are usually lowercase, so curly braces are used to preserve the correct casing. In NetLogo, as the name suggests, PascalCase is the norm. This convention is particularly useful when working with reference managers (e.g., Zotero, Mendeley), which often capitalize the first letter of a title or convert it to title case.

Related discussions:

- https://forum.posit.co/t/curly-braces-around-package-name-convention/122547
- https://www.reddit.com/r/rstats/comments/ep1bp9/why_do_people_write_package_names_between_curly/
