### 2025-09-18 #8845 @platipodium
> This is a tacit convention in the R community to distinguish package names in plain text (i.e., outside Markdown).

I am obviously not from the `R` community then... As an "outsider", it is quite unusual ... confusing.  And of course, LogoClim is *not* an R package so no reason to go with R conventions.
