### 2025-09-23 9c4103e
Remove curly braces from LogoClim title
