+++ _brand.yml
@@ -1,12 +1,12 @@
 meta:
   name:
-    short: "{LogoClim}"
-    long: "{LogoClim}: WorldClim in NetLogo"
+    short: "LogoClim"
+    long: "LogoClim: WorldClim in NetLogo"
   link:
     home: https://sustentarea.github.io/logoclim
     github: https://github.com/sustentarea/logoclim
   description: |
-    {LogoClim} is a NetLogo model for simulating and visualizing global climate
+    LogoClim is a NetLogo model for simulating and visualizing global climate
     conditions. It allows researchers to integrate high-resolution climate data
     into agent-based models, supporting reproducible research in ecology,
     agriculture, environmental sciences, and other fields that rely on climate
