### 2025-09-23 5f3580f
Remove the "Motivation" section from the JOSS paper
