### 2025-09-24 b8a408b
Add a color bar to the interface
