+++ qmd/data-munging.qmd
@@ -1,11 +1,3 @@
----
-params:
-  series: "historical-climate-data"
-  resolution: "10m"
-  model: NULL
-  country_code: "bra"
----
-
 # LogoClim: WorldClim 2.1 Data Transformation for NetLogo
 
 ## Overview
