+++ qmd/data-download.qmd
@@ -1,10 +1,3 @@
----
-params:
-  series: "historical-climate-data"
-  resolution: "10m"
-  model: NULL
----
-
 # LogoClim: WorldClim 2.1 Data Download
 
 ## Overview
