### 2025-10-29 #8845 @danielvartan
Hi @cheginit , @jamesdamillington , and @platipodium ,

I'm really sorry for the delay. I had planned to finish everything over the past few days, but an unexpected work emergency kept me from completing the remaining tasks.

I recently transitioned to a new fellowship, so things are still a bit unsettled. I’d like to ask for a little more time to wrap up the last issues. Once I return from a symposium on November 10th, I’ll finalize my responses to the pending points.
