### 2025-12-04 439325a
Update visual identity
