+++ R/_post-render.R
@@ -2,12 +2,12 @@ library(beepr)
 library(groomr) # github.com/sustentarea/groomr
 library(here)
 
-# Remove Empty Lines From `README.md` -----
+# Remove Empty Lines from `README.md` -----
 
 here("README.md") |> remove_blank_line_dups()
 
 # Check If the Script Ran Successfully -----
 
-beepr::beep(1)
+beep(1)
 
 Sys.sleep(3)
