+++ cloc.sh
@@ -1,1 +1,0 @@
-git ls-files | grep -e "\.nls" -e "\.nlogox" | xargs wc -l
