+++ ttf/README.md
@@ -1,0 +1,9 @@
+# LogoClim
+
+## Typefaces
+
+### Inter
+
+- License: Open Font License (v1.1): <https://openfontlicense.org/>
+- Source: Google Fonts
+- Downloaded on: 2025-09-25
