+++ README.qmd
@@ -7,17 +7,22 @@
 #| include: false
 
 library(badger)
-library(beepr)
-library(groomr) # github.com/danielvartan/groomr
-library(lubridate)
-library(readr)
-library(stringr)
+library(brandr)
+library(dplyr)
+library(ggplot2)
+library(here)
+library(htmltools)
+library(knitr)
+
+source(here("R", "_setup.R"))
 ```
 
 <!-- badges: start -->
 ```{r}
 #| echo: false
 #| output: asis
+#| message: false
+#| warning: false
 
 cat(
   paste0(
@@ -38,13 +43,15 @@ cat(
     url = paste0(
       "https://www.comses.net/codebases/",
       "bccd451f-76a4-408a-85fd-c5024359ba9a/"
-    )
+    ),
+    alt = "CoMSES Network Badge"
   ),
   badge_custom(
     x = "OSF DOI",
     y = "10.17605/OSF.IO/EAPZU",
     color = "1284C5",
-    url = "https://doi.org/10.17605/OSF.IO/EAPZU"
+    url = "https://doi.org/10.17605/OSF.IO/EAPZU",
+    alt = "OSF DOI Badge"
   ),
   paste0(
     "[![",
@@ -74,13 +81,15 @@ cat(
   badge_license(
     license = "GPLv3",
     color = "bd0000",
-    url = "https://www.gnu.org/licenses/gpl-3.0"
+    url = "https://www.gnu.org/licenses/gpl-3.0",
+    alt = "GNU GPLv3 License"
   ),
   badge_custom(
     x = "Contributor%20Covenant",
     y = "3.0",
     color = "4baaaa",
-    url = "https://www.contributor-covenant.org/version/3/0/code_of_conduct/"
+    url = "https://www.contributor-covenant.org/version/3/0/code_of_conduct/",
+    alt = "Contributor Covenant 3.0 Code of Conduct Badge"
   )
 )
 ```
@@ -96,6 +105,8 @@ See the [`Logônia`](https://github.com/sustentarea/logonia) model for an exampl
 
 > If you find this project useful, please consider giving it a star! &nbsp; [![GitHub repo stars](https://img.shields.io/github/stars/sustentarea/logoclim)](https://github.com/sustentarea/logoclim/)
 
+> The continuous development of `LogoClim` depends on community support. If you find this project useful, and can afford to do so, please consider becoming a sponsor. &nbsp; [![](https://img.shields.io/static/v1?label=Sponsor&message=%E2%9D%A4&logo=GitHub&color=%23fe8e86)](https://github.com/sponsors/danielvartan)
+
 ```{=html}
 <p align="center">
   <img src="images/logoclim-interface.gif" />
@@ -103,7 +114,7 @@ See the [`Logônia`](https://github.com/sustentarea/logonia) model for an exampl
 ```
 
 ```{=gfm}
-> `r "[!IMPORTANT]"`
+> [!IMPORTANT]
 > `LogoClim` is an independent project with no affiliation to [WorldClim](https://worldclim.org/) or its developers. Users should be aware that WorldClim datasets are freely available for academic and other non-commercial use only. Any use of WorldClim data within `LogoClim` must comply with [WorldClim's licensing terms](https://worldclim.org/about.html).
 ```
 
@@ -207,16 +218,21 @@ cat(
     x = "Contributor%20Covenant",
     y = "3.0",
     color = "4baaaa",
-    url = "https://www.contributor-covenant.org/version/3/0/code_of_conduct/"
+    url = "https://www.contributor-covenant.org/version/3/0/code_of_conduct/",
+    alt = "Contributor Covenant 3.0 Code of Conduct"
   )
 )
 ```
 
[…]

[…]
