### 2025-12-04 748f3ef
Add `check-netlogo` workflow
