### 2025-12-04 d14c452
Add unbundled NetLogo extensions
