+++ nlogox/extensions/pathdir/src/PathDirExtension.java
@@ -1,0 +1,594 @@
+// PathDirExtension - V 5.1.0 (August 2025)
+// updated for NetLogo v7.0
+// Charles Staelin, Smith College
+
+/*
+ * Contains a number of procedures for finding paths, for creating,
+ * renaming and deleting directories, and for finding file sizes and dates.
+ * REMEMBER THAT ANY PROCEDURES THAT FOOL WITH YOUR FILES MAY BE DANGEROUS!
+ */
+
+ /* updated for NetLogo v7.0.0.  Making pathdir.jar requires two other jar
+  * files: netlogo-7.x.x.jar (where the x's represent the subversion numbers
+  * of the current NetLogo installation) and scala-library-2.13.16.jar. Both 
+  * jar files are found in the /app subdirectoy of the NetLogo installation.
+ */
+package org.nlogo.extensions.pathdir;
+
+import org.nlogo.api.*;
+//import org.nlogo.core.LogoList;
+import org.nlogo.core.Syntax;
+import org.nlogo.core.SyntaxJ;
+import java.util.Arrays;
+import java.util.ArrayList;
+import java.util.Date;
+import java.util.Properties;
+import java.util.Enumeration;
+import java.util.Map;
+import java.text.SimpleDateFormat;
+import java.io.File;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.InputStreamReader;
+import java.io.BufferedReader;
+
+public class PathDirExtension extends org.nlogo.api.DefaultClassManager {
+
+    private static org.nlogo.workspace.ExtensionManager em;
+
+    @Override
+    // runOnce is no longer used, but left here in case it might be useful
+    // in the future. It gets and holds the current ExtensionManager.
+    public void runOnce(org.nlogo.api.ExtensionManager em)
+            throws org.nlogo.api.ExtensionException {
+        PathDirExtension.em = (org.nlogo.workspace.ExtensionManager) em;
+    }
+
+    // Prepends to the attachName argument the current working directory as
+    // specified in the NetLogo model's context.  However, if attachName is an
+    // absolute path, it is returned unchanged.
+    private static String attachCWD(Context context, String attachName)
+            throws ExtensionException {
+        if (attachName.length() == 0) {
+            attachName = ".";
+        }
+        try {
+            attachName = context.attachCurrentDirectory(attachName);
+        } catch (java.net.MalformedURLException ex) {
+            throw new ExtensionException(ex);
+        }
+
+        File f = new File(attachName);
+        try {
+            return (f.getCanonicalFile()).toString();
+        } catch (IOException ex) {
+            ExtensionException eex = new ExtensionException(ex);
+            eex.setStackTrace(ex.getStackTrace());
+            throw eex;
+        }
+    }
+
+    // strips the path and the file extension from the string.
+    // should be pretty robust with wierd paths and filenames.
+    private static String getFilenameOnly(String s) {
+
+        String separator = File.separator;
+        String filename;
+
+        // Remove the path upto the filename.
+        int lastSeparatorIndex = s.lastIndexOf(separator);
+        if (lastSeparatorIndex == -1) {
+            filename = s;
+        } else {
+            filename = s.substring(lastSeparatorIndex + 1);
+        }
+
+        // Remove the extension.
+        int extensionIndex = filename.lastIndexOf(".");
+        if (extensionIndex == -1) {
+            return filename;
+        }
+
+        return filename.substring(0, extensionIndex);
+    }
+
+    private static ProcessBuilder _pb;
+
+    private static void initializeProcessBuilder(Context context) {
+        if (_pb == null) {
+            _pb = new ProcessBuilder(new ArrayList<>());
+            _pb.redirectErrorStream(true);
[…]
