+++ nlogox/extensions/pathdir/README.md
@@ -1,0 +1,250 @@
+# NetLogo pathdir Extension
+Version 5.2 - September 2025 (for NetLogo v7)
+
+* [What is it?](#what-is-it)
+* [Installation](#installation)
+* [Examples](#examples)
+* [Primitives](#primitives)
+* [Author](#author)
+* [Feedback](#feedback-bugs-feature-requests)
+* [Credits](#credits)
+* [Terms of use](#terms-of-use)
+
+## What is it?
+
+This package contains the NetLogo **pathdir extension**, which provides NetLogo with some file-related primitives not included in the standard language, particularly primitives related to manipulating directories, identifying the current NetLogo model, moving files, finding the size and modification dates of files, and getting the values of environment variables and system properties.
+
+[back to top](#netlogo-pathdir-extension)
+
+## Installation
+
+Include the extension in your NetLogo model using the **extensions** primitive. It should be the first line in your model. (If there are already other extensions in your model, simply add *pathdir* to the list.)
+
+    extensions [pathdir]
+
+If the extension has not yet been installed, NetLogo will prompt you to install it and, upon your replying in the affirmative, will download the extension and its associated files to the appropriate location.
+
+For more information on NetLogo extensions:
+[http://ccl.northwestern.edu/netlogo/docs/extensions.html](http://ccl.northwestern.edu/netlogo/docs/extensions.html)
+
+[back to top](#netlogo-pathdir-extension)
+
+## Examples
+
+See the **PathDir.nlogox** model for examples of usage.
+
+
+[back to top](#netlogo-pathdir-extension)
+
+## Primitives
+
+**pathdir:get-separator**
+**pathdir:get-file-separator**
+
+*pathdir:get-separator*
+*pathdir:get-file-separator*
+
+Returns a string with the character used by the host operating system to separate directories in a given path.  E.g., for Windows the string "\\\" would be returned (as the backslash must be escaped), while for Mac OSX and linux, the string "/" would be returned.  Useful for creating operating system-independent paths to files.
+
+---------------------------------------
+
+**pathdir:get-path-separator**
+
+*pathdir:get-path-separator**
+
+Returns a string with the character used by the host operating system to separate paths in a list of paths, such as in Windows "PATH" environmental variable or in java.library.path. 
+E.g., for Windows the string ";" would be returned while for Mac OSX and linux, the string ":" would be returned.  Useful for creating operating system-independent lists of paths.
+
+---------------------------------------
+
+**pathdir:get-model-path**
+
+*pathdir:get-model-path*
+
+Returns a string with the full (absolute) path to the directory in which the current model is located.
+
+NOTE: Returns an empty string ("") if the current model has not yet been saved to a file.
+
+---------------------------------------
+
+**pathdir:get-model-file**
+
+*pathdir:get-model-file*
+
+Returns a string with the filename of the .nlogox or .nlogo3dx file containing the current model. 
+
+NOTE: Returns an empty string ("") if the current model has not yet been saved to a file.
+
+Stripping the .nlogox extension  (or any extension) is easily done:
+
+    let modelName pathdir:get-model-name
+    let shortName substring modelName 0 (length modelName - position "." reverse modelName - 1)
+(Note that the code above can handle filenames with embedded periods.)
+
+The model filename can also be concatenated with the path to it:
+
+    let fullModelPath (word pathdir:get-model pathdir:get-separator pathdir:get-model-name)
+
+---------------------------------------
+**pathdir:get-model-name**
+
+*pathdir:get-model-name*
+
+Returns a string with the name of the current model. NetLogo sets the name of the model to the filename of the .nlogox or .nlogo3dx file containing the model, stripped of the file extension.
+
[…]
