### 2025-12-04 b20d20a
Add experiments that perform parameter sweeps with check-netlogo
