### 2026-01-11 538a738
Fix NetLogo check for Windows runners
