### 2026-01-12 a65a479
Add near equality tests
