### 2026-01-12 172794e
Add draft user manual
