### 2026-01-12 #8845 @danielvartan
Hi everyone,

Here are some updates.

The [`logolink`](https://danielvartan.github.io/logolink/index.html) R package v1.0.0 has been officially released on CRAN. This release resolves one of the main blockers for the new `LogoClim` tests. I didn't consider it safe to fully automate testing while relying on the previous version.

[`LogoActions`](https://github.com/danielvartan/logoactions) v1.4.0 has also been released, bringing significant improvements to the `setup-netlogo` and `check-netlogo` GitHub Actions. These now support Windows and macOS runners in addition to Linux, enabling to test nearly all `LogoClim` parameter combinations across the three main operating systems. These checks continue to run on every commit, but now across all three platforms.

You can access the latest parameter test logs here: https://github.com/sustentarea/logoclim/actions/runs/20919731668

I've implemented the first version of the near-equality and procedures tests, along with a draft of the User Manual.

The near-equality tests compare the original WorldClim GeoTIFF data against values extracted directly from NetLogo patches using randomly selected countries and parameter combinations. The comparison applies a relative tolerance of 1%, and only areas with at least 1,000 cells are included to ensure statistical reliability.

Both the near-equality tests and the procedure behavior checks are integrated into the User Manual appendices and will be rendered automatically with each commit. While they could run separately, integrating them into the manual makes inspection and auditing more straightforward.

You can preview the draft near-equality test report here: https://sustentarea.github.io/logoclim/qmd/near-equality-tests.html

Please note that both the tests and user manual are still in draft form and not yet ready for detailed review. I'm continuing work on the final responses.
