### 2026-01-12 adcd287
Remove `render-manual.yaml`
