+++ .github/workflows/render-manual.yaml
@@ -1,95 +1,0 @@
-on:
-  push:
-    branches: [main, master]
-
-name: render-manual.yaml
-permissions: read-all
-
-jobs:
-  render-manual:
-    name: Render user manual
-    runs-on: ubuntu-latest
-    env:
-      GH_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-    permissions:
-      contents: write
-      pages: write
-    steps:
-      - name: Check out repository
-        uses: actions/checkout@v4
-
-      - name: Set up NetLogo
-        uses: danielvartan/logoactions/setup-netlogo@v1
-
-      - name: Set up R
-        uses: r-lib/actions/setup-r@v2
-
-      - name: Set up Quarto
-        uses: quarto-dev/quarto-actions/setup@v2
-
-      - name: Install system dependencies
-        if: runner.os == 'Linux'
-        run: |
-          # Install system dependencies
-
-          sudo apt-get update -qq
-
-          sudo apt-get install -y -qq \
-            libfontconfig1-dev \
-            pandoc
-        shell: bash
-
-      - name: Install R dependencies
-        run: |
-          # Install R dependencies
-
-          options(repos=c(CRAN="https://cloud.r-project.org"))
-
-          install.packages(
-            c(
-              "rmarkdown",
-              "knitr"
-            )
-          )
-        shell: Rscript {0}
-
-      - name: Check if renv is initialized
-        id: renv-check
-        run: |
-          # Check if renv is initialized
-
-          if [ -f "renv.lock" ]; then
-            echo "exists=true" >> $GITHUB_OUTPUT
-          else
-            echo "exists=false" >> $GITHUB_OUTPUT
-          fi
-        shell: bash
-
-      - name: Set up renv
-        if: steps.renv-check.outputs.exists == 'true'
-        uses: r-lib/actions/setup-renv@v2
-
-      - name: Install and initialize renv
-        if: steps.renv-check.outputs.exists == 'false'
-        run: |
-          # Install and initialize renv
-
-          install.packages("renv")
-
-          renv::init()
-        shell: Rscript {0}
-        working-directory: './tests/qmd'
-
-      - name: Render Quarto
-        run: |
-          # Render Quarto
-
-          quarto render --profile html
-        shell: bash
-
-      - name: Deploy to GitHub Pages
-        uses: JamesIves/github-pages-deploy-action@v4.5.0
-        with:
-          clean: false
-          branch: gh-pages
-          folder: docs
