### 2026-02-06 #8845 @jamesdamillington
Hi All, [issue #9](https://github.com/sustentarea/logoclim/issues/9) is still open, and from the last message of @danielvartan I inferred I should wait to check the new tests that have been added? @danielvartan let me know if you think I can go ahead and check the test functionality.
