### 2026-02-18 0254b92
Remove docs and improve near equality tests
