+++ R/compare_plots.R
@@ -1,82 +1,151 @@
 library(checkmate)
-library(dplyr)
 library(here)
-library(magrittr)
+library(ggplot2)
 library(patchwork)
-library(terra)
 
-source(here("R", ".setup.R"))
+here("R", ".setup.R") |> source()
+here("R", "worldclim_raster.R") |> source()
+here("R", "logoclim_raster.R") |> source()
 
+#' Plot and compare WorldClim GeoTIFF and LogoClim patch values
+#'
+#' @description
+#'
+#' `compare_plots()` generates comparative plots for
+#' [WorldClim](https://www.worldclim.org)
+#' [GeoTIFF](https://en.wikipedia.org/wiki/GeoTIFF) values and
+#' [LogoClim](https://github.com/sustentarea/logoclim) patch values. It
+#' visualizes the spatial distribution of climate data from WorldClim and the
+#' corresponding patch values from LogoClim, allowing for a visual comparison
+#' between the two datasets.
+#'
+#' See LogoClim's [`near-equality-tests.qmd`](
+#' https://github.com/sustentarea/logoclim/tree/main/qmd) Quarto notebook for
+#' examples of how to use this function.
+#'
+#' @param tif_file A [`character`][base::character()] string specifying the path
+#'   to the [WorldClim](https://www.worldclim.org)
+#'   [GeoTIFF](https://en.wikipedia.org/wiki/GeoTIFF) file.
+#' @param country_shape A [`SpatVector`][terra::SpatVector()] object
+#'   representing a country's spatial extent for cropping the
+#'   [WorldClim](https://www.worldclim.org) data, as returned by the
+#'   [`gadm()`][geodata::gadm()] function from the
+#'   [`geodata`](https://rspatial.github.io/geodata) package.
+#' @param results A [`list`][base::list()] containing the results from the
+#'   [LogoClim](https://github.com/sustentarea/logoclim) model, as output by
+#'   the
+#'   [`logolink`](https://danielvartan.github.io/logolink) R package, containing
+#'   the [NetLogo](https://www.netlogo.org)
+#'   [BehaviorSpace](https://docs.netlogo.org/behaviorspace) output formats
+#'   [`table`](https://docs.netlogo.org/behaviorspace#table-output)
+#'   and
+#'   [`lists`](https://docs.netlogo.org/behaviorspace#lists-output).
+#' @param setup A [`list`][base::list()] containing the setup information of
+#'   the comparison, as provided by the
+#'   [`worldclim_random()`][orbis::worldclim_random()].
+#' @param layer_pattern (optional) A [`character`][base::character()] string
+#'   specifying the layer pattern to use when subsetting from a multi-layer
+#'   [WorldClim](https://www.worldclim.org)
+#'   [GeoTIFF](https://en.wikipedia.org/wiki/GeoTIFF) file. When `NULL`, the
+#'   first layer is used (default: `NULL`).
+#' @param dx (optional) A [`numeric`][base::numeric()] value specifying the
+#'   degree of longitude shift to apply when handling spatial data that crosses
+#'   the date line (default: `-45`). See the
+#'   [`shift_and_rotate()`][orbis::shift_and_rotate()] function from the
+#'   [`orbis`](https://danielvartan.github.io/orbis) R package for more details
+#'   on this process.
+#' @param remove_extreme_outliers (optional) A [`logical`][base::logical()] flag
+#'   indicating whether to transform to `NA` values 10 times the interquartile
[…]
