+++ docs/robots.txt
@@ -1,1 +1,0 @@
-Sitemap: https://sustentarea.github.io/logoclim/sitemap.xml

[…]
