### 2026-02-18 e4d87bc
Remove non .qmd files from qmd directory
