+++ .github/workflows/netlogo-check.yaml
@@ -7,6 +7,7 @@ permissions: read-all
 
 jobs:
   check-netlogo:
+    if: "!contains(github.event.head_commit.message, '[skip netlogo check]')"
     strategy:
       fail-fast: false
       matrix:
