### 2026-02-18 efd5fcc
Add system dependencies to render-manual.yaml [skip netlogo check]
