+++ qmd/near-equality-tests.qmd
@@ -156,6 +156,7 @@ library(fs)
 library(geodata)
 library(ggplot2)
 library(here)
+library(ISOcodes)
 library(knitr)
 library(leaflet)
 library(logolink)
