+++ .github/workflows/render-manual.yaml
@@ -18,15 +18,6 @@ jobs:
       - name: Check out repository
         uses: actions/checkout@v4
 
-      - name: Set up NetLogo
-        uses: danielvartan/logoactions/setup-netlogo@v1
-
-      - name: Set up R
-        uses: r-lib/actions/setup-r@v2
-
-      - name: Set up Quarto
-        uses: quarto-dev/quarto-actions/setup@v2
-
       - name: Install system dependencies
         if: runner.os == 'Linux'
         run: |
@@ -45,6 +36,15 @@ jobs:
             libx11-dev \
             pandoc
         shell: bash
+
+      - name: Set up NetLogo
+        uses: danielvartan/logoactions/setup-netlogo@v1
+
+      - name: Set up R
+        uses: r-lib/actions/setup-r@v2
+
+      - name: Set up Quarto
+        uses: quarto-dev/quarto-actions/setup@v2
 
       - name: Install R dependencies
         run: |
