### 2026-02-18 389acda
Add system dependencies to render-manual.yaml [skip netlogo check]
