### 2026-02-18 45cc205
Add system dependencies to render-manual.yaml [skip netlogo check]
