### 2026-02-18 9db4bcb
Add system dependencies to render-manual.yaml [skip netlogo check]
