+++ .github/workflows/render-manual.yaml
@@ -32,7 +32,7 @@ jobs:
             libcairo2-dev \
             libcurl4-openssl-dev \
             libfontconfig1-dev \
-            libfribidi-dev
+            libfribidi-dev \
             libgdal-dev \
             libgit2-dev \
             libharfbuzz-dev \
