### 2026-02-18 b4fb839
Fix .pre-render.R libraries [skip netlogo check]
