+++ .github/workflows/render-manual.yaml
@@ -45,11 +45,11 @@ jobs:
       - name: Set up NetLogo
         uses: danielvartan/logoactions/setup-netlogo@v1
 
-      - name: Set up R
-        uses: r-lib/actions/setup-r@v2
-
       - name: Set up Quarto
         uses: quarto-dev/quarto-actions/setup@v2
+
+      - name: Set up R
+        uses: r-lib/actions/setup-r@v2
 
       - name: Install R dependencies
         run: |
