+++ robots.txt
@@ -1,0 +1,1 @@
+Sitemap: https://sustentarea.github.io/logoclim/sitemap.xml
