### 2026-02-19 6a02feb
Improve user manual [skip netlogo check]
