+++ qmd/how-to-use-it.qmd
@@ -9,61 +9,67 @@ library(here)
 here("R", ".setup.R") |> source()
 ```
 
-::: {.callout-important}
-This manual is still under development and may be subject to change.
+To get started using `LogoClim`, you must have [NetLogo](https://www.netlogo.org) version 7 or later installed. The NetLogo [website](https://www.netlogo.org) provides easy installers for Windows, macOS, and Linux, along with detailed instructions for installation.
 
-This warning will be removed once the manual is finalized.
+The model also depends on four NetLogo extensions: [`GIS`](https://docs.netlogo.org/gis.html), [`Pathdir`](https://github.com/cstaelin/Pathdir-Extension), [`String`](https://github.com/NetLogo/String-Extension), and [`Time`](https://docs.netlogo.org/time.html). No manual installation is required since they are automatically installed the first time the model runs.
+
+::: {.callout-tip}
+Linux users can install NetLogo via [`LogoPak`](https://github.com/danielvartan/logopak), a [Flatpak](https://flatpak.org) package that bundles all four NetLogo applications: NetLogo, NetLogo 3D, HubNet Client, and BehaviorSearch.
 :::
 
-To get started using `LogoClim`, ensure you have [NetLogo](https://www.netlogo.org) installed. The current version of the model was developed using NetLogo 7.0.3, so it is recommended to use this version or later.
+With NetLogo ready, follow these 5 steps to get `LogoClim` up and running.
 
-`LogoClim` also relies on the [`GIS`](https://docs.netlogo.org/gis.html), [`Pathdir`](https://github.com/cstaelin/Pathdir-Extension), [`String`](https://github.com/NetLogo/String-Extension), and [`Time`](https://docs.netlogo.org/time.html) NetLogo extensions. You don't need to install them, these are automatically installed when the model is run for the first time.
+## A. Downloading the Model
 
-## Downloading the Model
+You can download the latest release of the model from the [CoMSES Network](https://www.comses.net/codebases/bccd451f-76a4-408a-85fd-c5024359ba9a/). This is the recommended option for most users, as it provides a stable version of the model that has been tested and documented.
 
-You can download the latest release of the model from the [CoMSES Network](https://www.comses.net/codebases/bccd451f-76a4-408a-85fd-c5024359ba9a/) or at its [GitHub releases page](https://github.com/sustentarea/logoclim/releases/latest).
+For the development version, you can clone or download the model [GitHub code repository](https://github.com/sustentarea/logoclim/) directly.
 
-For the development version, you can clone or download the [GitHub code repository](https://github.com/sustentarea/logoclim/) directly.
+## B. Opening the Model
 
-## Opening the Model
+After downloading and uncompressing the model files, open the `logoclim.nlogox` file in NetLogo. You can find this file in the `code` directory when using the [CoMSES Network](https://www.comses.net/codebases/bccd451f-76a4-408a-85fd-c5024359ba9a/) release.
 
-After downloading and uncompressing the model files, open the `logoclim.nlogox` file in NetLogo. This file is located in the [`nlogox`](nlogox) folder.
+## C. Preparing the Data
 
-## Preparing the Data
+The [CoMSES Network release](https://www.comses.net/codebases/bccd451f-76a4-408a-85fd-c5024359ba9a/) includes an example dataset that is ready to use with `LogoClim`. You can use it as a starting point. But, ideally you should prepare your own data to suit your research needs. The next sections of this manual will guide you through the process of downloading and preparing [WorldClim](https://www.worldclim.org/) data for use with `LogoClim`.
 
[…]
