### 2026-02-19 sustentarea/logoclim#17 @danielvartan
Hi @platipodium,

As mentioned in https://github.com/sustentarea/logoclim/issues/14, I have reworked all the procedures related to downloading and converting the data, and documented everything in the `LogoClim` [user manual](https://sustentarea.github.io/logoclim/) (see chapters 1 to 5). I also followed your suggestion and set [`-99`](https://danielvartan.github.io/orbis/reference/worldclim_to_ascii.html#:~:text=overwrite%20=%20TRUE%2C-,na_flag%20=%20%2D99%2C,-dir%20=%20dirname%28file%5B1%5D%29) as the default value for the `NODATA` field.

I believe this addresses what you were looking for. Please let me know if anything is missing or if you would like me to go further on any point.
