### 2026-02-19 #8845 @danielvartan
Hi @jamesdamillington, @platipodium, and @cheginit,

Thank you for all your patience, and sorry for the delay.

I have finished all the unit tests, including the [near-equality tests](https://sustentarea.github.io/logoclim/qmd/near-equality-tests.html) and the [procedure behavior tests](https://sustentarea.github.io/logoclim/qmd/procedure-behavior-tests.html). I describe them in detail in [issue #9](https://github.com/sustentarea/logoclim/issues/9).

I also reworked all the procedures related to downloading and converting data, and documented everything in the `LogoClim` [user manual](https://sustentarea.github.io/logoclim/) (see chapters 1 to 5).

I believe this addresses the three remaining open issues:

- [Issue #9](https://github.com/sustentarea/logoclim/issues/9)
- [Issue #14](https://github.com/sustentarea/logoclim/issues/14)
- [Issue #17](https://github.com/sustentarea/logoclim/issues/17)

@jamesdamillington @platipodium, please let me know if anything is missing or if you would like me to go further on any point.
