### 2026-02-19 d1b052e
Update user manual
