### 2026-02-21 sustentarea/logoclim#9 @danielvartan
Along with the Continuous Integration tests described above, the [`Logônia`](https://github.com/sustentarea/logonia) example model also helps verify the functionality of `LogoClim`.

![Image](https://github.com/user-attachments/assets/a3ab0d92-e156-4246-b95f-5e38957d44f2)

![Image](https://github.com/user-attachments/assets/f7fb04f1-b0d2-4f7b-ab6c-d14fce1c1fb4)

> Variation of Average Minimum Temperature (°C) (A), Average Maximum Temperature (°C) (B), and Total Precipitation (mm) (C) from the [WorldClim 2.1](https://worldclim.org/) Historical Monthly Weather Data series for a region of the Brazilian Amazon. Each frame represents one month, showing how climate conditions shift over time. The [Logônia](https://github.com/sustentarea/logonia) model uses this data to simulate how plants respond to changing climate conditions.

![Image](https://github.com/user-attachments/assets/7eb41c84-d4d4-409b-83d7-84ecbc300d8e)

![Image](https://github.com/user-attachments/assets/6b93a115-5176-494a-8edc-75edc50becbe)

![Image](https://github.com/user-attachments/assets/659e804e-65ac-4e15-b0b9-fa0ffba3e639)
