### 2026-02-22 f22ee41
Add wrap to NEWS.md in release-notes.qmd
