### 2026-02-22 #8845 @jamesdamillington
@danielvartan   This all looks great and addresses my concerns for JOSS. I have been able to run the near-equality-tests and procedure-behaviour-tests out of the box on ubuntu (installing for Windows had a hitch which I highlight on [Issue #9](https://github.com/sustentarea/logoclim/issues/9) but as the tests are verified on ubuntu I don't see this as a problem for the JOSS paper). See my comment also on [Issue #14](https://github.com/sustentarea/logoclim/issues/14). I did not have a problem with [Issue #17](https://github.com/sustentarea/logoclim/issues/17). Congratulations on a nice piece of software that I'm sure many will find useful and as @cheginit notes, thanks for working so diligently in addressing my comments.

@cheginit I have checked off all items in my checklist and from my perspective this is good to go for JOSS
