### 2026-02-22 41356cc
Update renv dependencies and improve the documentation
