+++ README.qmd
@@ -14,7 +14,7 @@ library(here)
 library(htmltools)
 library(knitr)
 
-source(here("R", ".setup.R"))
+here("R", ".setup.R") |> source()
 ```
 
 <!-- badges: start -->
@@ -36,16 +36,6 @@ cat(
     ")"
   ),
   badge_repostatus("active"),
-  badge_github_actions(
-    ref = "sustentarea/logoclim",
-    action = "check-netlogo.yaml",
-    alt = "Check NetLogo workflow badge"
-  ),
-  badge_github_actions(
-    ref = "sustentarea/logoclim",
-    action = "render-manual.yaml",
-    alt = "Render manual workflow badge"
-  ),
   badge_custom(
     x = "CoMSES Network",
     y = "2.1.0",
@@ -63,17 +53,15 @@ cat(
     url = "https://doi.org/10.17605/OSF.IO/EAPZU",
     alt = "OSF DOI badge"
   ),
-  paste0(
-    "[![",
-    "FAIR checklist badge",
-    "](",
-    "https://img.shields.io/badge/",
-    "fairsoftwarechecklist.net", "-",
-    "-",
-    "00a7d9",
-    ")](",
-    "https://fairsoftwarechecklist.net/v0.2?f=31&a=30112&i=32301&r=123",
-    ")"
+  badge_github_actions(
+    ref = "sustentarea/logoclim",
+    action = "check-netlogo.yaml",
+    alt = "Check NetLogo workflow badge"
+  ),
+  badge_github_actions(
+    ref = "sustentarea/logoclim",
+    action = "render-manual.yaml",
+    alt = "Render manual workflow badge"
   ),
   paste0(
     "[![",
@@ -86,6 +74,18 @@ cat(
     "green",
     ")](",
     "https://fair-software.eu",
+    ")"
+  ),
+  paste0(
+    "[![",
+    "FAIR checklist badge",
+    "](",
+    "https://img.shields.io/badge/",
+    "fairsoftwarechecklist.net", "-",
+    "-",
+    "00a7d9",
+    ")](",
+    "https://fairsoftwarechecklist.net/v0.2?f=31&a=30112&i=32301&r=123",
     ")"
   ),
   badge_license(
@@ -150,7 +150,7 @@ This series includes 12 monthly data points from [downscaled](https://worldclim.
 
 Refer to the `Info` tab in the model for additional details.
 
-## How to Use It
+## Usage
 
 ### Setup
 
@@ -192,63 +192,26 @@ See the [`Logônia`](https://github.com/sustentarea/logonia) model for an exampl
 
 ## User Manual
 
-`LogoClim`'s user manual was developed using the latest versions of the [Quarto](https://quarto.org/) publishing system, the [NetLogo](https://www.netlogo.org/) environment, and the [R](https://www.r-project.org/) programming language. To ensure consistent results, the [`renv`](https://rstudio.github.io/renv/) package was used to manage and restore the R environment.
+`LogoClim`'s user manual is developed using the latest versions of the [Quarto](https://quarto.org/) publishing system, the [NetLogo](https://www.netlogo.org/) environment, and the [R](https://www.r-project.org/) programming language. To ensure consistent results, the [`renv`](https://rstudio.github.io/renv/) package is used to manage and restore the R environment.
 
 To reproduce the analyses in the manual, install the four dependencies mentioned above and follow these steps:
 
 1. **Clone** this repository to your local machine.
 2. **Open** the project in your preferred [IDE](https://en.wikipedia.org/wiki/Integrated_development_environment).
-3. **Restore the R environment** by running [`renv::restore()`](https://rstudio.github.io/renv/reference/restore.html) in the R console. This will install all required software dependencies.
[…]
