+++ qmd/release-notes.qmd
@@ -1,4 +1,4 @@
-# Developer Notes {#sec-developer-notes}
+# Release Notes {#sec-release-notes}
 
 ```{r}
 #| label: setup
@@ -16,10 +16,10 @@ here("R", ".setup.R") |> source()
 ## Changelog
 
 <!-- %:::% NEWS.md begin %:::% -->
-``` {.txt filename="NEWS.md"}
-# v2.1.0.9000 (development version)
+### v2.1.0.9000 (development version)
 
-- A first draft of `LogoClim` User Manual is now available in the `docs/` folder and online at [https://sustentarea.github.io/logoclim/](https://sustentarea.github.io/logoclim/). The manual is a work in progress and will be continuously updated.
+- A first draft of `LogoClim` User Manual is now available in the `docs` folder and online at [https://sustentarea.github.io/logoclim/](https://sustentarea.github.io/logoclim/). The manual is a work in progress and will be continuously updated.
+- Automated unit tests on near-equality comparisons between WorldClim original files and `LogoClim` patch values, as well as procedure behavior tests, are now included in the user manual and rendered at each commit.
 - A `color-bar-bins` slider was added to control the number of bins in the color bar.
 - The interface now has a new color bar widget for better visualization of patch values.
 - The *Bioclimatic Variable* monitor was removed. Now it is displayed directly on the *Climate Variable* monitor.
@@ -31,7 +31,7 @@ here("R", ".setup.R") |> source()
 - Code of Conduct updated to [Contributor Covenant 3.0](https://www.contributor-covenant.org/version/3/0/code_of_conduct/).
 - The documentation was updated to reflect all changes.
 
-# v2.1.0 (2025-09-16)
+### v2.1.0 (2025-09-16)
 
 - A `#headless?` parameter was added to the `setup` procedure for improved headless execution.
 - The `#tick` and `#wait` parameters were removed from the `go` procedure; it now has no parameters, simplifying execution.
@@ -46,13 +46,13 @@ here("R", ".setup.R") |> source()
 - [Quarto](https://quarto.org/) notebooks were updated.
 - The documentation was updated to reflect all changes.
 
-# v2.0.0 (2025-07-29)
+### v2.0.0 (2025-07-29)
 
-### Breaking Changes
+#### Breaking Changes
 
 - The model now requires **NetLogo 7.0.1*- or later.
 
-### New Features and Improvements
+#### New Features and Improvements
 
 - Interface widgets were resized to match the new NetLogo standard dimensions.
 - The model's license was changed from [MIT](https://opensource.org/license/mit) to [GPLv3](https://opensource.org/license/gpl-3-0).
@@ -60,16 +60,16 @@ here("R", ".setup.R") |> source()
 - The `adjust-world-size?` slider was removed; world size adjustment is now always enabled by default.
 - A global variable named `plot-max-y-range` was introduced to optimize computations.
 
-# v1.0.0 (2025-07-03)
+### v1.0.0 (2025-07-03)
 
 First stable release! 🎉
 
-# v0.0.0.9015 (2025-07-02) (Pre-Release)
+### v0.0.0.9015 (2025-07-02) (Pre-Release)
 
 - Variations (e.g., ACCESS-ESM1-5) and additional (e.g., CanESM5) Global Climate Models (GCMs) were added as selectable options in `global-climate-model`. WorldClim provides a dedicated webpage for this data, available [here](https://www.worldclim.org/data/cmip6_all/cmip6_clim2.5m.html).
 - The year interval issue when using the *Future Climate Data- series was fixed.
 
[…]

[…]
