### 2026-02-22 sustentarea/logoclim#9 @danielvartan
Hi @jamesdamillington,

Thank you for your review!

`renv` can be [a little frustrating](https://youtu.be/l01u7Ue9pIQ?si=S44LlHVSufGJ4zdq) at times. I added instructions on how to deactivate it in the [README](https://github.com/sustentarea/logoclim#user-manual) to avoid issues like the one you described.

I also updated all R packages in my machine and created a new `renv` dependency file with the latest version of `ggiraph`. It should work on Windows now.
