### 2026-02-23 #8845 @danielvartan
Thank you for the review and kind words, @jamesdamillington!

I updated the repository, refreshing the `renv` dependencies and adding instructions in the [README](https://github.com/sustentarea/logoclim#user-manual) on how to remove the `renv` environment when needed. It should work on Windows now.
