### 2026-02-23 sustentarea/logoclim#9 @danielvartan
Closing this issue to make the [JOSS review](https://github.com/openjournals/joss-reviews/issues/8845) process easier to follow.
