### 2026-02-23 sustentarea/logoclim#9 @jamesdamillington
Hi @danielvartan  On Windows, I re-cloned the latest version of the repo this morning into a directory for which the path had no spaces - this (plus maybe updated package versions) overcame the package installation issues on Windows. However, when running the near-equality-tests on Windows, the logoclim `run_experiment` fails (see below). I pulled the same version of the repo to my ubuntu installation and everything worked fine. I have NetLogo 7.0.3 installed on Windows and it seems to work fine via the GUI (e.g. to test, I downloaded data for France from the OSF repository).

```
Error in `run_experiment()`:
! NetLogo produced the following error during the experiment run:

Picked up JAVA_TOOL_OPTIONS: -Dcom.sun.media.jai.disableMediaLib=true
java.util.NoSuchElementException: head of empty list
Child process exited with code 1
Run `rlang::last_trace()` to see where the error occurred.

> rlang::last_trace(drop = FALSE)
<error/rlang_error>
Error in `run_experiment()`:
! NetLogo produced the following error during the experiment run:

Picked up JAVA_TOOL_OPTIONS: -Dcom.sun.media.jai.disableMediaLib=true
java.util.NoSuchElementException: head of empty list
Child process exited with code 1
---
Backtrace:
    ▆
 1. └─logolink::run_experiment(model_path, setup_file = setup_file, output = c("table", "lists"))
 2.   └─cli::cli_abort(...)
 3.     └─rlang::abort(...)
```
