### 2026-02-23 0f39071
Add Windows and macOS runners in render-manual.yaml
