+++ .github/workflows/check-netlogo.yaml
@@ -7,7 +7,7 @@ permissions: read-all
 
 jobs:
   check-netlogo:
-    if: "!contains(github.event.head_commit.message, '[skip netlogo check]')"
+    if: "!contains(github.event.head_commit.message, '[skip check-netlogo]')"
     strategy:
       fail-fast: false
       matrix:
