### 2026-02-23 e3565f5
Add labels and snapshot to render-manual [skip check-netlogo]
