+++ .github/workflows/render-manual.yaml
@@ -59,20 +59,6 @@ jobs:
       - name: Set up R
         uses: r-lib/actions/setup-r@v2
 
-      - name: Install R dependencies
-        run: |
-          # Install R dependencies
-
-          options(repos=c(CRAN="https://cloud.r-project.org"))
-
-          install.packages(
-            c(
-              "rmarkdown",
-              "knitr"
-            )
-          )
-        shell: Rscript {0}
-
       - name: Check if renv is initialized
         id: renv-check
         run: |
@@ -94,9 +80,27 @@ jobs:
         run: |
           # Install and initialize renv
 
+          options(repos=c(CRAN="https://cloud.r-project.org"))
+
           install.packages("renv")
 
           renv::init()
+        shell: Rscript {0}
+
+      - name: Install R dependencies
+        run: |
+          # Install R dependencies
+
+          options(repos=c(CRAN="https://cloud.r-project.org"))
+
+          install.packages(
+            c(
+              "rmarkdown",
+              "knitr"
+            )
+          )
+
+          renv::snapshot()
         shell: Rscript {0}
 
       - name: Render Quarto
