### 2026-02-23 sustentarea/logoclim#9 @jamesdamillington
@danielvartan  I'm afraid it failed again (19:05 and 19:11 GMT). Can you suggest a way to quickly edit code in near-equality-tests.qmd to use the data I downloaded from OSF instead of a random country from WorldClim server, just to prove this is the issue? I inspected the temporary asc created right before the `run_experiment` is called and it looks okay? (see below, random country was SSD, `setup` shown below)

```
$series
Historical Climate Data
                  "hcd"

$resolution
10 Minutes (~340 km2 at the Equator)
                               "10m"

$variable
Wind Speed (m s^-1)
             "wind"

$year
1970-2000
     1994

$month
December
      12
```
