### 2026-02-24 9fea99f
Add workaround  in near-equality tests for Windows paths [skip check-netlogo]
