### 2026-02-24 281add8
Add macOS fix to logoclim_raster()
