### 2026-02-24 1bd1fc6
Add print_setup() to near-equality tests [skip check-netlogo]
