### 2026-02-24 sustentarea/logoclim#9 @danielvartan
Found it: another issue related to path incompatibility between Windows and Unix-based systems.

When running BehaviorSpace experiments in headless mode, NetLogo does not accept single backslashes in values. You need to escape them (e.g., `\\`) or, for file paths, use forward slashes instead. XML handles single backslashes just fine, but Java does not. This is especially relevant for Windows users, since Windows uses backslashes in file paths by default.

I opened an issue on the NetLogo repository and updated the `logolink` documentation to warn users about it.

https://github.com/NetLogo/NetLogo/issues/2638

The fix is straightforward: just add `|> normalizePath("/")` after `tempdir()`.
```r
"data-path" = tempdir() |> normalizePath("/")
```

I already applied this in the `LogoClim` repository. I also removed the developer version of the `s2` package from the `renv` dependencies, since it required system-level installations and made the setup error-prone.

The CI now renders the user manual and the tests on Windows, macOS, and Ubuntu runners. Here is the latest run:

https://github.com/sustentarea/logoclim/actions/runs/22333803016/job/64621767080

Ubuntu and Windows runs are passing. I'm still investigating the issue with the macOS runner.
