### 2026-02-24 4a22e51
Update print_setup()
