+++ R/print_setup.R
@@ -25,15 +25,17 @@ print_setup <- function(setup, title = "Selected WorldClim dataset") {
   assert_list(setup, min.len = 1)
   assert_string(title)
 
-  cli_h2(title)
+  message <- paste0(title, ":\n\n")
+
+  cli::cli_alert_info(message)
 
   for (i in names(setup)) {
     value <- setup[[i]]
     label <- names(value)
     value <- unname(value)
 
-    cli_bullets(
-      c("*" = paste0("{.field {i}}: {.val {value}} ({label})"))
+    cli_alert(
+      paste0("{.field {i}}: {.val {value}} ({label})")
     )
   }
 
