### 2026-02-24 c0ab62c
Update info messages in the near-equality tests [skip check-netlogo]
