### 2026-02-24 ee78bde
Update info messages in the near-equality tests [skip check-netlogo]
