### 2026-02-24 d1da587
Update README [skip check-netlogo] [skip render-manual]
