### 2026-02-24 sustentarea/logoclim#14 @platipodium
Also add that `libgit2` as a dependency.  Wow ... renv is taking > 1 hour (no, really, just 46 minutes)
 so far on my Mac M5 :=(
