+++ README.md
@@ -27,37 +27,18 @@ badge](https://img.shields.io/badge/Contributor%20Covenant-3.0-4baaaa.svg)](http
 
 ## Overview
 
-`LogoClim` is a [NetLogo](https://www.netlogo.org) model for simulating
-and visualizing global climate conditions. It allows researchers to
-integrate high-resolution climate data into agent-based models,
-supporting reproducible research in ecology, agriculture, environmental
-sciences, and other fields that rely on climate data.
+`LogoClim` is a [NetLogo](https://www.netlogo.org) model designed to
+simulate and visualize global climate conditions. It allows researchers
+to pull high-resolution climate data directly into agent-based models,
+making it easier to study how climate variables interact with complex
+systems over time.
 
-The model utilizes raster data to represent climate variables such as
-temperature and precipitation over time. It incorporates historical data
-(1951-2024) and future climate projections (2021-2100) derived from
-[global climate
-models](https://www.climatehubs.usda.gov/hubs/northwest/topic/basics-global-climate-models)
-under various Shared Socioeconomic Pathways
-([SSPs](https://climatedata.ca/resource/understanding-shared-socio-economic-pathways-ssps/),
-[O’Neill et al.,
-2017](https://doi.org/10.1016/j.gloenvcha.2015.01.004)). All climate
-inputs come from [WorldClim 2.1](https://worldclim.org/), a widely used
-source of high-resolution, interpolated climate datasets based on
-weather station observations worldwide ([Fick & Hijmans,
-2017](https://doi.org/10.1002/joc.5086)).
-
-See the [`Logônia`](https://github.com/sustentarea/logonia) model for an
-example of how to integrate `LogoClim` into your model.
+Learn more about the model in the [user
+manual](https://sustentarea.github.io/logoclim/).
 
 > If you find this project useful, please consider giving it a star!  
 > [![GitHub repo
 > stars](https://img.shields.io/github/stars/sustentarea/logoclim)](https://github.com/sustentarea/logoclim/)
-
-> The continuous development of `LogoClim` depends on community support.
-> If you find this project useful, and can afford to do so, please
-> consider becoming a sponsor.  
-> [![](https://img.shields.io/static/v1?label=Sponsor&message=%E2%9D%A4&logo=GitHub&color=%23fe8e86)](https://github.com/sponsors/danielvartan)
 
 <p align="center">
   <img src="images/logoclim-interface.gif" />
@@ -68,35 +49,45 @@ example of how to integrate `LogoClim` into your model.
 
 ## How It Works
 
-`LogoClim` operates on a grid of patches, where each patch represents a
+`LogoClim` uses raster data to represent climate variables such as
+temperature and precipitation over time. It incorporates historical data
+(1951-2024) and future climate projections (2021-2100) derived from
+[global climate
+models](https://www.climatehubs.usda.gov/hubs/northwest/topic/basics-global-climate-models)
+under various Shared Socioeconomic Pathways
+([SSPs](https://climatedata.ca/resource/understanding-shared-socio-economic-pathways-ssps/),
+[O’Neill et al.,
+2017](https://doi.org/10.1016/j.gloenvcha.2015.01.004)).
+
+The model operates on a grid of patches, where each patch represents a
 geographical area and stores values for latitude, longitude, and
 selected climate variables. During the simulation, patches update their
 colors based on the data values. The results can be visualized on a map,
 accompanied by plots that display the mean, minimum, maximum, and
 standard deviation of the selected variable over time.
 
-### Data Series
+All climate inputs come from [WorldClim 2.1](https://worldclim.org/), a
+widely used source of high-resolution climate datasets based on weather
+station observations worldwide ([Fick & Hijmans,
[…]
