### 2026-02-24 sustentarea/logoclim#14 @danielvartan
> Makes me wonder whether `renv` is such a smart move after all, as its installation procedure is tedious and might throw off users. I was much quicker before manually resolving the R dependencies ...

`renv` is not required to use `LogoClim`. It's only needed if you want to render the user manual or run the unit tests. For data download and processing, only the `orbis` package is required, as described in the user manual:

https://sustentarea.github.io/logoclim/qmd/data-download.html
