### 2026-02-24 sustentarea/logoclim#14 @danielvartan
About `s2`: try pulling the latest state of the `LogoCliim`repository. I removed the developer version of `s2` as a dependency and switched to the stable release.
