### 2026-02-24 sustentarea/logoclim#14 @platipodium
> https://sustentarea.github.io/logoclim/qmd/data-processing.html

Yes, that worked!
