### 2026-02-25 675c191
Test macOS runner [skip check-netlogo]
