+++ qmd/near-equality-tests.qmd
@@ -518,7 +518,11 @@ results |> glimpse()
 
 cli_alert_info("Experiment results:\n\n")
 
-results |> glimpse()
+results |>
+  glimpse() |>
+  capture.output() |>
+  paste(collapse = "\n") |>
+  cli_alert_info()
 ```
 
 ### Compare Plots
