### 2026-02-25 5230325
Install system dependencies on macOS in render-manual.yaml [skip check-netlogo]
