+++ .github/workflows/render-manual.yaml
@@ -26,10 +26,25 @@ jobs:
       - name: Check out repository
         uses: actions/checkout@v4
 
-      - name: Install system dependencies
+      - name: Install system dependencies (macOS)
+        if: runner.os == 'macOS'
+        run: |
+          # Install system dependencies (macOS)
+
+          brew install pkg-config
+
+          brew install \
+            gdal \
+            proj \
+            geos
+
+          echo "PROJ_LIB=/opt/homebrew/share/proj" >> "${GITHUB_ENV}"
+        shell: bash
+
+      - name: Install system dependencies (Linux)
         if: runner.os == 'Linux'
         run: |
-          # Install system dependencies
+          # Install system dependencies (Linux)
 
           sudo apt-get update -qq
 
@@ -102,6 +117,14 @@ jobs:
 
           renv::snapshot()
         shell: Rscript {0}
+
+      - name: Check GDAL/PROJ Config (macOS)
+        if: runner.os == 'macOS'
+        run: |
+          # Check GDAL/PROJ Config (macOS)
+
+          Rscript -e "terra::gdal(lib = 'all')"
+        shell: bash
 
       - name: Render Quarto
         run: |
