### 2026-02-25 4c7e4cf
Use path_norm() for resolving paths in the near-equality tests
