+++ README.qmd
@@ -191,7 +191,7 @@ To facilitate this integration, we created the [`Logônia`](https://github.com/s
 
 ```{=html}
 <p align="center">
-  <img src="images/logonia-evolution.svg" height=125px />
+  <img src="images/logonia-interface.gif" />
 </p>
 ```
 
