### 2026-02-26 c321336
Use local temp directory in near-equality tests [skip check-netlogo]
