### 2026-02-26 ad95fb5
Remove system dependencies installation in render-manual.yaml
