+++ .github/workflows/render-manual.yaml
@@ -26,45 +26,6 @@ jobs:
       - name: Check out repository
         uses: actions/checkout@v4
 
-      - name: Install system dependencies (macOS)
-        if: runner.os == 'macOS'
-        run: |
-          # Install system dependencies (macOS)
-
-          brew install pkg-config
-
-          brew install \
-            gdal \
-            proj \
-            geos
-
-          echo "PROJ_LIB=/opt/homebrew/share/proj" >> "${GITHUB_ENV}"
-        shell: bash
-
-      - name: Install system dependencies (Linux)
-        if: runner.os == 'Linux'
-        run: |
-          # Install system dependencies (Linux)
-
-          sudo apt-get update -qq
-
-          sudo apt-get install -y -qq \
-            cmake \
-            gdal-bin \
-            gsfonts \
-            libcairo2-dev \
-            libcurl4-openssl-dev \
-            libfontconfig1-dev \
-            libfribidi-dev \
-            libgdal-dev \
-            libgit2-dev \
-            libharfbuzz-dev \
-            libmagick++-dev \
-            libudunits2-dev \
-            libx11-dev \
-            pandoc
-        shell: bash
-
       - name: Set up NetLogo
         uses: danielvartan/logoactions/setup-netlogo@v1
 
@@ -74,57 +35,8 @@ jobs:
       - name: Set up R
         uses: r-lib/actions/setup-r@v2
 
-      - name: Check if renv is initialized
-        id: renv-check
-        run: |
-          # Check if renv is initialized
-
-          if [ -f "renv.lock" ]; then
-            echo "exists=true" >> $GITHUB_OUTPUT
-          else
-            echo "exists=false" >> $GITHUB_OUTPUT
-          fi
-        shell: bash
-
       - name: Set up renv
-        if: steps.renv-check.outputs.exists == 'true'
         uses: r-lib/actions/setup-renv@v2
-
-      - name: Install and initialize renv
-        if: steps.renv-check.outputs.exists == 'false'
-        run: |
-          # Install and initialize renv
-
-          options(repos=c(CRAN="https://cloud.r-project.org"))
-
-          install.packages("renv")
-
-          renv::init()
-        shell: Rscript {0}
-
-      - name: Install R dependencies
-        run: |
-          # Install R dependencies
-
-          options(repos=c(CRAN="https://cloud.r-project.org"))
-
-          install.packages(
-            c(
-              "rmarkdown",
-              "knitr"
-            )
-          )
-
-          renv::snapshot()
-        shell: Rscript {0}
-
-      - name: Check GDAL/PROJ Config (macOS)
-        if: runner.os == 'macOS'
-        run: |
-          # Check GDAL/PROJ Config (macOS)
-
-          Rscript -e "terra::gdal(lib = 'all')"
-        shell: bash
 
       - name: Render Quarto
         run: |
