### 2026-02-26 5cea1cd
Clean cache in render-manual.yaml and enable ppm in renv [skip check-netlogo]
