### 2026-02-26 b00df4d
Add system dependencies to render-manual.yaml [skip check-netlogo]
