+++ README.qmd
@@ -197,24 +197,29 @@ To facilitate this integration, we created the [`Logônia`](https://github.com/s
 
 ## User Manual
 
+```{=gfm}
+> [!NOTE]
+> This section describes the technical setup required to render the user manual locally. You do not need any of this to simply use `LogoClim`.
+```
+
 `LogoClim`'s [user manual](https://sustentarea.github.io/logoclim/) is developed using the latest versions of the [Quarto](https://quarto.org/) publishing system, the [NetLogo](https://www.netlogo.org/) environment, and the [R](https://www.r-project.org/) programming language. To ensure consistent results, the [`renv`](https://rstudio.github.io/renv/) package is used to manage and restore the R environment.
 
-To render the manual or reproduce its analyses locally, install the four dependencies listed above and follow the steps below. **You do not need any of this to use `LogoClim`**, this is only relevant if you want to contribute to its development or run the unit tests yourself.
+To render the manual or reproduce its analyses locally, install the four dependencies listed above and follow the steps below.
 
 1. **Clone** this repository to your local machine.
 2. **Open** the project in the terminal or in your preferred [IDE](https://en.wikipedia.org/wiki/Integrated_development_environment).
 3. **Install package dependencies** by running `Rscript -e "renv::restore()"` in the terminal or [`renv::restore()`](https://rstudio.github.io/renv/reference/restore.html) in an R console. Make sure R is installed and available in your system's [PATH](https://en.wikipedia.org/wiki/PATH_(variable)) before running this command.
 4. **Open** the Quarto notebook files (`.qmd`) and run the code as described.
 
-We do not recommend using external environments such as [Anaconda](https://www.anaconda.com/), as these can cause issues with R package installation and management. This project relies on several system dependencies like [GDAL](https://gdal.org/), all of which are automatically installed via the Comprehensive R Archive Network ([CRAN](https://cran.r-project.org/)).
+When running [`renv::restore()`](https://rstudio.github.io/renv/reference/restore.html), check the output for any missing system dependencies like [GDAL](https://gdal.org/). They are usually installed automatically via your OS package manager, but if something fails, you may need to handle them manually. See [`render-manual.yaml`](.github/workflows/render-manual.yaml) for a list of system dependencies required for your operating system.
 
-If you have trouble installing R, we recommend using the installers provided by the [R Project](https://www.r-project.org/) or the [`rig`](https://github.com/r-lib/rig) installation manager from [`r-lib`](https://github.com/r-lib). If you are using an [IDE](https://en.wikipedia.org/wiki/Integrated_development_environment) with no built-in R console, we also recommend installing the [`radian`](https://github.com/randy3k/radian) console for a better experience.
+We do not recommend using external environments such as [Anaconda](https://www.anaconda.com/), as these can cause issues with R package installation and management. This project relies on several system dependencies, all of which are automatically installed via the Comprehensive R Archive Network ([CRAN](https://cran.r-project.org/)).
+
+We recommend using the installers provided by the [R Project](https://www.r-project.org/) or the [`rig`](https://github.com/r-lib/rig) installation manager from [`r-lib`](https://github.com/r-lib) when installing R. If your [IDE](https://en.wikipedia.org/wiki/Integrated_development_environment) lacks a built-in R console, consider installing [`radian`](https://github.com/randy3k/radian) for a better experience.
 
 Avoid using [VPNs](https://en.wikipedia.org/wiki/Virtual_private_network), corporate proxies, or other network-routing tools while processing the data, as these can interfere with the downloads.
 
[…]
