### 2026-02-26 220ec45
Change render-manual.yaml to install renv before Quarto
