+++ .github/workflows/render-manual.yaml
@@ -68,9 +68,6 @@ jobs:
       - name: Set up NetLogo
         uses: danielvartan/logoactions/setup-netlogo@v1
 
-      - name: Set up Quarto
-        uses: quarto-dev/quarto-actions/setup@v2
-
       - name: Set up R
         uses: r-lib/actions/setup-r@v2
 
@@ -78,6 +75,9 @@ jobs:
         uses: r-lib/actions/setup-renv@v2
         with:
           cache-version: 1 # default: 1
+
+      - name: Set up Quarto
+        uses: quarto-dev/quarto-actions/setup@v2
 
       - name: Render Quarto
         run: |
