+++ .github/workflows/render-manual.yaml
@@ -74,7 +74,7 @@ jobs:
       - name: Set up renv
         uses: r-lib/actions/setup-renv@v2
         with:
-          cache-version: 1 # default: 1
+          cache-version: 2 # default: 1
 
       - name: Set up Quarto
         uses: quarto-dev/quarto-actions/setup@v2
