+++ .github/workflows/render-manual.yaml
@@ -68,6 +68,9 @@ jobs:
       - name: Set up NetLogo
         uses: danielvartan/logoactions/setup-netlogo@v1
 
+      - name: Set up Quarto
+        uses: quarto-dev/quarto-actions/setup@v2
+
       - name: Set up R
         uses: r-lib/actions/setup-r@v2
 
@@ -75,9 +78,6 @@ jobs:
         uses: r-lib/actions/setup-renv@v2
         with:
           cache-version: 2 # default: 1
-
-      - name: Set up Quarto
-        uses: quarto-dev/quarto-actions/setup@v2
 
       - name: Render Quarto
         run: |
