### 2026-02-26 22b95d9
Check renv::status in render-manual.yaml [skip check-netlogo]
