+++ references.bib
@@ -46,7 +46,7 @@
   date = {2014},
   publisher = {O'Reilly Media},
   location = {Sebastopol, CA},
-  url = {https://rstudio-education.github.io/hopr/},
+  url = {https://rstudio-education.github.io/hopr},
   abstract = {Learn how to program by diving into the R language, and then use your newfound skills to solve practical data science problems. With this book, you’ll learn how to load data, assemble and disassemble data objects, navigate R’s environment system, write your own functions, and use all of R’s programming tools. RStudio Master Instructor Garrett Grolemund not only teaches you how to program, but also shows you how to get more from R than just visualizing and modeling data. You’ll gain valuable programming skills and support your work as a data scientist at the same time.},
   isbn = {978-1-4493-5901-0},
   langid = {american},
@@ -283,7 +283,7 @@
   author = {Wilensky, Uri},
   date = {1999},
   location = {Evanston, IL},
-  url = {http://ccl.northwestern.edu/netlogo/},
+  url = {https://www.netlogo.org},
   organization = {Center for Connected Learning and Computer-Based Modeling, Northwestern University},
   keywords = {agent-based modeling,complexity science,computer science,exact sciences,interdisciplinary fields,netlogo,programming}
 }
