### 2026-02-26 sustentarea/logoclim#14 @platipodium
I cloned fresh, installed R, quarto again, ran renv, then:

```
(logoclim) ➜  logoclim git:(main) quarto render --profile html
R/.pre-render.R
# Bootstrapping renv 1.1.7 ---------------------------------------------------
- Downloading renv ... OK
- Installing renv  ... OK

- One or more packages recorded in the lockfile are not installed.
- Use `renv::status()` for more details.
Error in library(checkmate) : there is no package called ‘checkmate’
Execution halted
```
