### 2026-02-26 cb705dc
Change near-equality tests to report the setup experiments
