### 2026-02-26 4676edc
Refresh renv cache in render-manual.yaml
