+++ .github/workflows/render-manual.yaml
@@ -77,7 +77,7 @@ jobs:
       - name: Set up renv
         uses: r-lib/actions/setup-renv@v2
         with:
-          cache-version: 1 # default: 1
+          cache-version: 2 # default: 1
 
       - name: Render Quarto
         run: |
