### 2026-02-26 ca1a797
Remove renv cache refresh in render-manual.yaml [skip check-netlogo]
