+++ .github/workflows/render-manual.yaml
@@ -31,8 +31,6 @@ jobs:
         run: |
           # Install system dependencies (macOS)
 
-          brew install pkg-config
-
           brew install \
             gdal \
             proj \
@@ -77,7 +75,7 @@ jobs:
       - name: Set up renv
         uses: r-lib/actions/setup-renv@v2
         with:
-          cache-version: 2 # default: 1
+          cache-version: 1 # default: 1
 
       - name: Render Quarto
         run: |
