### 2026-02-26 ec60462
Fix leaflet visualization in near-equality tests
