### 2026-02-27 sustentarea/logoclim#14 @platipodium
```
Rscript -e "renv::restore()"
...
- Downloading fs 1.6.6 from CRAN ...            OK [file is up to date]
- Downloading rappdirs 0.3.4 from CRAN ...      OK [file is up to date]
- Downloading desc 1.4.3 from CRAN ...          OK [file is up to date]
...
- Downloading groomr 0.1.0.9000 from GitHub ... - GitHub authentication credentials are not available.
- Please set GITHUB_PAT, or ensure the 'gitcreds' package is installed.
- See https://usethis.r-lib.org/articles/git-credentials.html for more details.
OK [169.4 Kb in 0.52s]
...
- Installing digest 0.6.39 ...                  OK [linked from cache]
- Installing fs 1.6.6 ...                       FAILED
Error: Error installing package 'fs':
==============================

* installing *source* package ‘fs’ ...
** this is package ‘fs’ version ‘1.6.6’
** package ‘fs’ successfully unpacked and MD5 sums checked
** using staged installation
** libs
using C++ compiler: ‘Apple clang version 17.0.0 (clang-1700.6.3.2)’
using SDK: ‘MacOSX26.2.sdk’
clang++ -std=gnu++17 -I"/opt/homebrew/Cellar/r/4.5.2_1/lib/R/include" -DNDEBUG -I./libuv-1.44.2/include -I. -pthread  -I/opt/homebrew/opt/gettext/include -I/opt/homebrew/opt/readline/include -I/opt/homebrew/opt/xz/include -I/opt/homebrew/include    -fPIC  -g -O2   -c dir.cc -o dir.o
clang++ -std=gnu++17 -I"/opt/homebrew/Cellar/r/4.5.2_1/lib/R/include" -DNDEBUG -I./libuv-1.44.2/include -I. -pthread  -I/opt/homebrew/opt/gettext/include -I/opt/homebrew/opt/readline/include -I/opt/homebrew/opt/xz/include -I/opt/homebrew/include    -fPIC  -g -O2   -c error.cc -o error.o
clang++ -std=gnu++17 -I"/opt/homebrew/Cellar/r/4.5.2_1/lib/R/include" -DNDEBUG -I./libuv-1.44.2/include -I. -pthread  -I/opt/homebrew/opt/gettext/include -I/opt/homebrew/opt/readline/include -I/opt/homebrew/opt/xz/include -I/opt/homebrew/include    -fPIC  -g -O2   -c file.cc -o file.o
clang++ -std=gnu++17 -I"/opt/homebrew/Cellar/r/4.5.2_1/lib/R/include" -DNDEBUG -I./libuv-1.44.2/include -I. -pthread  -I/opt/homebrew/opt/gettext/include -I/opt/homebrew/opt/readline/include -I/opt/homebrew/opt/xz/include -I/opt/homebrew/include    -fPIC  -g -O2   -c fs.cc -o fs.o
clang++ -std=gnu++17 -I"/opt/homebrew/Cellar/r/4.5.2_1/lib/R/include" -DNDEBUG -I./libuv-1.44.2/include -I. -pthread  -I/opt/homebrew/opt/gettext/include -I/opt/homebrew/opt/readline/include -I/opt/homebrew/opt/xz/include -I/opt/homebrew/include    -fPIC  -g -O2   -c getmode.cc -o getmode.o
clang++ -std=gnu++17 -I"/opt/homebrew/Cellar/r/4.5.2_1/lib/R/include" -DNDEBUG -I./libuv-1.44.2/include -I. -pthread  -I/opt/homebrew/opt/gettext/include -I/opt/homebrew/opt/readline/include -I/opt/homebrew/opt/xz/include -I/opt/homebrew/include    -fPIC  -g -O2   -c id.cc -o id.o
clang++ -std=gnu++17 -I"/opt/homebrew/Cellar/r/4.5.2_1/lib/R/include" -DNDEBUG -I./libuv-1.44.2/include -I. -pthread  -I/opt/homebrew/opt/gettext/include -I/opt/homebrew/opt/readline/include -I/opt/homebrew/opt/xz/include -I/opt/homebrew/include    -fPIC  -g -O2   -c init.cc -o init.o
clang++ -std=gnu++17 -I"/opt/homebrew/Cellar/r/4.5.2_1/lib/R/include" -DNDEBUG -I./libuv-1.44.2/include -I. -pthread  -I/opt/homebrew/opt/gettext/include -I/opt/homebrew/opt/readline/include -I/opt/homebrew/opt/xz/include -I/opt/homebrew/include    -fPIC  -g -O2   -c link.cc -o link.o
clang++ -std=gnu++17 -I"/opt/homebrew/Cellar/r/4.5.2_1/lib/R/include" -DNDEBUG -I./libuv-1.44.2/include -I. -pthread  -I/opt/homebrew/opt/gettext/include -I/opt/homebrew/opt/readline/include -I/opt/homebrew/opt/xz/include -I/opt/homebrew/include    -fPIC  -g -O2   -c path.cc -o path.o
clang++ -std=gnu++17 -I"/opt/homebrew/Cellar/r/4.5.2_1/lib/R/include" -DNDEBUG -I./libuv-1.44.2/include -I. -pthread  -I/opt/homebrew/opt/gettext/include -I/opt/homebrew/opt/readline/include -I/opt/homebrew/opt/xz/include -I/opt/homebrew/include    -fPIC  -g -O2   -c utils.cc -o utils.o
clang++ -std=gnu++17 -I"/opt/homebrew/Cellar/r/4.5.2_1/lib/R/include" -DNDEBUG -I./libuv-1.44.2/include -I. -pthread  -I/opt/homebrew/opt/gettext/include -I/opt/homebrew/opt/readline/include -I/opt/homebrew/opt/xz/include -I/opt/homebrew/include    -fPIC  -g -O2   -c unix/getmode.cc -o unix/getmode.o
touch libuv-1.44.2/aclocal.m4 && touch libuv-1.44.2/configure && touch libuv-1.44.2/Makefile.in
(cd libuv-1.44.2 \
	&& CC="clang -std=gnu23" CPPFLAGS="-I/opt/homebrew/opt/gettext/include -I/opt/homebrew/opt/readline/include -I/opt/homebrew/opt/xz/include -I/opt/homebrew/include" CFLAGS="-g -O2  -fPIC -fvisibility=hidden -std=c99" AR="ar" RANLIB="ranlib" LDFLAGS="-L/opt/homebrew/opt/gettext/lib -L/opt/homebrew/opt/readline/lib -L/opt/homebrew/opt/xz/lib -L/opt/homebrew/lib" ./configure  --quiet)
rm: conftest.dSYM: is a directory
ld: warning: -single_module is obsolete
configure: error: in `/private/var/folders/p6/tn93l5qs7s307208rld23w_80000gp/T/RtmpUdrjqK/renv-package-new-86c516c867b6/fs/src/libuv-1.44.2':
configure: error: C preprocessor "arm64-apple-darwin20.0.0-clang-cpp" fails sanity check
See `config.log' for more details
make: *** [libuv-1.44.2/Makefile] Error 1
ERROR: compilation failed for package ‘fs’
* removing ‘/Users/Lemmen/devel/review/logoclim/renv/staging/1/fs’
install of package 'fs' failed [error code 1]
Traceback (most recent calls last):
13: renv::restore()
12: renv_restore_run_actions(project, diff, current, lockfile, rebuild)
11: renv_install_impl(records)
10: renv_install_staged(records)
 9: renv_install_default(records)
 8: handler(package, renv_install_package(record))
 7: renv_install_package(record)
 6: withCallingHandlers(renv_install_package_impl(record), error = function(e) writef("FAILED"))
 5: renv_install_package_impl(record)
 4: r_cmd_install(package, path)
 3: r_exec_error(package, output, "install", status)
 2: abort(all)
 1: stop(fallback)
Execution halted
```
