### 2026-02-27 sustentarea/logoclim#14 @danielvartan
Your log shows a mismatch:
```
configure: error: C preprocessor "arm64-apple-darwin20.0.0-clang-cpp" fails sanity check
```

This looks like a conflict between the R version installed via Homebrew and your actual system settings. This project relies on several system dependencies and, when run through third-party R installations, compilation steps like this one tend to fail.

As a first fix, try installing `fs` outside `renv` by opening a terminal and running:

```bash
Rscript -e "install.packages('fs', type = 'binary')"
```

After that, run `renv::restore()` again from the project root. If `fs` installed successfully, `renv` will link to it and skip the compilation step.

If that does not work, please install R using the official CRAN `.pkg` installer or via [`rig`](https://github.com/r-lib/rig) from the `r-lib` project, then run `renv::restore()` again. The documentation already advises against third-party R installers, and I will add a specific note about Homebrew R.
