### 2026-02-28 sustentarea/logoclim#14 @platipodium
> This looks like a conflict between the R version installed via Homebrew and your actual system settings. This project relies on several system dependencies and, when run through third-party R installations, compilation steps like this one tend to fail.

Yep, that's the reason for using home-brew, conda, or macports.

> Rscript -e "install.packages('fs', type = 'binary')"
> After that, run `renv::restore()` again from the project root. If `fs` installed successfully, `renv` will link to it and skip the

Error in install.packages("fs", type = "binary") :
  type 'binary' is not supported on this platform
Execution halted
