### 2026-02-28 sustentarea/logoclim#14 @platipodium
Rig worked to an extend: it got `fs` installed.  But then it failed compilation of `curl`.

I tend to give up at this point, and not build the documentation on Mac.  Sorry.
