### 2026-02-28 #8845 @platipodium
@cheginit I have checked all items in my checklist, and I recommend this to be published in JOSS.

There are still items pending building documentation on a specific system, but that should not hinder going forward with this publication.
