### 2026-02-28 sustentarea/logoclim#14 @danielvartan
I understand.

Thank you for all your help!
