### 2026-04-12 #8845 @cheginit
@danielvartan Sorry for the delay and I appreciate your patience. In the reference section of the paper the items on lines 85, 103, and 119 are missing journal or publisher. Please add them to the `bib` file. Once you're ready, release a new version containing all the changes made to address issues raised by the reviewers. Then create an archive on website similar to Zenodo, and provide me with the version and the generated DOI. Please make sure that the version, title, authors, and their ORCID in the generated archive match exactly your submission.
