### 2026-07-06 08c6c89
docs: update JOSS paper references [skip check-netlogo] [skip render-manual]
