+++ paper/paper.bib
@@ -9,15 +9,13 @@
   publisher = {Nature Publishing Group},
   issn = {1476-4687},
   doi = {10.1038/533452a},
-  url = {https://www.nature.com/articles/533452a},
-  urldate = {2024-06-09},
   abstract = {Survey sheds light on the "crisis" rocking research.},
   langid = {english},
   keywords = {reproducibility,science}
 }
 
 @article{barker2022,
-  title = {Introducing the {{FAIR Principles}} for research software},
+  title = {Introducing the FAIR Principles for research software},
   author = {Barker, Michelle and Chue Hong, Neil P. and Katz, Daniel S. and Lamprecht, Anna-Lena and Martinez-Ortiz, Carlos and Psomopoulos, Fotis and Harrow, Jennifer and Castro, Leyla Jael and Gruenpeter, Morane and Martinez, Paula Andrea and Honeyman, Tom},
   date = {2022-10-14},
   journaltitle = {Scientific Data},
@@ -28,8 +26,6 @@
   publisher = {Nature Publishing Group},
   issn = {2052-4463},
   doi = {10.1038/s41597-022-01710-x},
-  url = {https://www.nature.com/articles/s41597-022-01710-x},
-  urldate = {2025-04-15},
   abstract = {Research software is a fundamental and vital part of research, yet significant challenges to discoverability, productivity, quality, reproducibility, and sustainability exist. Improving the practice of scholarship is a common goal of the open science, open source, and FAIR (Findable, Accessible, Interoperable and Reusable) communities and research software is now being understood as a type of digital object to which FAIR should be applied. This emergence reflects a maturation of the research community to better understand the crucial role of FAIR research software in maximising research value. The FAIR for Research Software (FAIR4RS) Working Group has adapted the FAIR Guiding Principles to create the FAIR Principles for Research Software (FAIR4RS Principles). The contents and context of the FAIR4RS Principles are summarised here to provide the basis for discussion of their adoption. Examples of implementation by organizations are provided to share information on how to maximise the value of research outputs, and to encourage others to amplify the importance and impact of this work.},
   langid = {english},
   keywords = {fair principles,reproducibility,research management,science}
@@ -45,17 +41,16 @@
   pages = {106003},
   issn = {1364-8152},
   doi = {10.1016/j.envsoft.2024.106003},
-  url = {https://www.sciencedirect.com/science/article/pii/S1364815224000641},
-  urldate = {2025-08-20},
[…]
