### 2026-07-06 367c5d6
docs: add JOSS draft paper workflow [skip check-netlogo] [skip render-manual]
