### 2026-07-06 2622d29
docs: update JOSS paper references [skip check-netlogo] [skip render-manual]
