+++ paper/paper.bib
@@ -110,9 +110,9 @@
   shorttitle = {{{geodata}}},
   author = {Hijmans, Robert J. and Barbosa, Márcia and Ghosh, Aniruddha and Mandel, Alex},
   date = {2024},
-  location = {CRAN},
   doi = {10.32614/CRAN.package.geodata},
   abstract = {Functions for downloading of geographic data for use in spatial analysis and mapping. The package facilitates access to climate, crops, elevation, land use, soil, species occurrence, accessibility, administrative boundaries and other data.},
+  organization = {CRAN},
   keywords = {apis,computer science,exact sciences,nosource,probability and statistics,r (programming language),r packages,software engineering,spatial data,spatial data analysis}
 }
 
@@ -182,8 +182,8 @@
   shorttitle = {Logônia},
   author = {Vartanian, Daniel and Garcia, Leandro and Carvalho, Aline Martins},
   date = {2025},
-  location = {GitHub},
   url = {https://github.com/sustentarea/logonia},
   abstract = {Logônia is a NetLogo model that simulates the growth response of a fictional plant, logônia, under different climatic conditions. The model uses climate data from WorldClim 2.1 and demonstrates how to integrate the LogoClim model through the LevelSpace extension.},
+  organization = {GitHub},
   keywords = {climate change,computer science,environmental sciences,exact sciences,interdisciplinary fields,logoclim,netlogo,nosource,probability and statistics,worldclim}
 }
