+++ README.qmd
@@ -1,470 +1,0 @@
-# LogoClim <img src="images/logo.svg" align="right" width="120" />
-
-<!-- quarto render -->
-
-```{r}
-#| label: setup
-#| include: false
-
-library(badger)
-library(brandr)
-library(dplyr)
-library(ggplot2)
-library(here)
-library(htmltools)
-library(knitr)
-
-here("R", ".setup.R") |> source()
-```
-
-<!-- badges: start -->
-```{r}
-#| echo: false
-#| output: asis
-#| message: false
-#| warning: false
-
-cat(
-  paste0(
-    "[![",
-    "JOSS journal badge",
-    "](",
-    "https://joss.theoj.org/papers/",
-    "093e7333c6a34b5f3dc0b44b76e05480/status.svg",
-    ")](",
-    "https://joss.theoj.org/papers/093e7333c6a34b5f3dc0b44b76e05480",
-    ")"
-  ),
-  badge_repostatus("active"),
-  badge_custom(
-    x = "CoMSES Network",
-    y = "2.1.0",
-    color = "1284C5",
-    url = paste0(
-      "https://www.comses.net/codebases/",
-      "bccd451f-76a4-408a-85fd-c5024359ba9a/"
-    ),
-    alt = "CoMSES Network badge"
-  ),
-  badge_custom(
-    x = "doi",
-    y = "10.17605/OSF.IO/EAPZU",
-    color = "1284C5",
-    url = "https://doi.org/10.17605/OSF.IO/EAPZU",
-    alt = "OSF DOI badge"
-  ),
-  badge_github_actions(
-    ref = "sustentarea/logoclim",
-    action = "check-netlogo.yaml",
-    alt = "Check NetLogo workflow badge"
-  ),
-  badge_github_actions(
-    ref = "sustentarea/logoclim",
-    action = "render-manual.yaml",
-    alt = "Render manual workflow badge"
-  ),
-  paste0(
-    "[![",
-    "fair-software.eu",
-    "](",
-    "https://img.shields.io/badge/",
-    "fair--software.eu", "-",
-    "%E2%97%8F%20%20%E2%97%8F%20%20%E2%97%8F%20%20%E2%97%8F%20%20%E2%97%8F",
-    "-",
-    "green",
-    ")](",
-    "https://fair-software.eu",
-    ")"
-  ),
-  paste0(
-    "[![",
-    "FAIR checklist badge",
-    "](",
-    "https://img.shields.io/badge/",
-    "fairsoftwarechecklist.net", "-",
-    "-",
-    "00a7d9",
-    ")](",
-    "https://fairsoftwarechecklist.net/v0.2?f=31&a=30112&i=32301&r=123",
-    ")"
-  ),
-  badge_license(
-    license = "GPLv3",
-    color = "bd0000",
-    url = "https://www.gnu.org/licenses/gpl-3.0",
-    alt = "GNU GPLv3 license"
-  ),
-  badge_custom(
-    x = "Contributor%20Covenant",
-    y = "3.0",
-    color = "4baaaa",
-    url = "https://www.contributor-covenant.org/version/3/0/code_of_conduct/",
-    alt = "Contributor Covenant 3.0 code of conduct badge"
-  )
-)
-```
-<!-- badges: end -->
-
-## Overview
-
-`LogoClim` is a [NetLogo](https://www.netlogo.org) model designed to simulate and visualize global climate conditions. It allows researchers to pull high-resolution climate data directly into agent-based models, making it easier to study how climate variables interact with complex systems over time.
-
-Learn more about the model in the [user manual](https://sustentarea.github.io/logoclim/).
-
[…]

[…]
