### 2026-07-11 e24b9a5
docs: update documentation [skip check-netlogo] [skip render-manual]
