+++ CHANGELOG.md
@@ -1,35 +1,20 @@
-# Release Notes {#sec-release-notes}
+# Changelog
 
-```{r}
-#| label: setup
-#| include: false
+All notable changes to this project will be documented in this file.
 
-library(here)
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-here("R", ".setup.R") |> source()
-```
+## v2.1.0.9000 (development version)
 
-## Notes
-
-- Based on Seth Tisue's [suggestion](https://github.com/NetLogo/NetLogo/issues/2554), `LogoClim` represents missing values as `false`. This resolves issues with primitives like `export-world` and `import-world` (see [GIS Known Issues](https://docs.netlogo.org/gis.html#known-issues)).
-
-## Changelog
-
-<!-- %:::% CHANGELOG.md begin %:::% -->
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
-and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-
-### v2.1.0.9000 (development version)
-
-#### Added
+### Added
 
 - A first draft of `LogoClim` User Manual, now available at [https://sustentarea.github.io/logoclim/](https://sustentarea.github.io/logoclim/). The manual is a work in progress and will be continuously updated.
 - A new color bar widget for better visualization of patch values.
 - A new global variable `cell-size` to store the size of each patch in degrees.
-- Automated unit tests covering most supported configurations using the [`check-netlogo`](https://github.com/danielvartan/netlogo-actions) action from the [`LogoActions`](https://github.com/danielvartan/logoactions) project. Tests run on Windows, macOS, and Linux with the latest NetLogo release at each commit. Due to computational constraints, only 10m resolution settings are tested. For future climate data, only SSP-126 is included because other SSPs are not consistently available across the Global Climate Models (GCMs). [GCM variations](https://www.worldclim.org/data/cmip6/cmip6_clim10m.html#:~:text=GCMs%20there%20are-,variations,-available%20here.) are not tested.
+- Automated unit tests covering most supported configurations using the [`check-netlogo`](https://github.com/danielvartan/netlogo-actions) action from the [`LogoActions`](https://github.com/danielvartan/logoactions) project. Tests run on Windows, macOS, and Linux with the latest NetLogo release at each commit. Due to computational constraints, only `10m` resolution settings are tested. For future climate data, only SSP-126 is included because other SSPs are not consistently available across the Global Climate Models (GCMs). [GCM variations](https://www.worldclim.org/data/cmip6/cmip6_clim10m.html#:~:text=GCMs%20there%20are-,variations,-available%20here.) are not tested.
 - Automated unit tests on near-equality comparisons between WorldClim original files and `LogoClim` patch values, as well as procedure behavior tests, now included in the user manual and rendered at each commit. Tests run on Windows, macOS, and Linux with the latest NetLogo release.
 
-#### Changed
+### Changed
 
 - The *Bioclimatic Variable* monitor was removed and is now displayed directly on the *Climate Variable* monitor.
 - Slider limits were adjusted and value validation for blank/empty settings was added.
@@ -37,21 +22,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - All dependencies were updated to their latest versions.
[…]
