### 2026-07-11 8b5804f
docs: add minor tweaks to JOSS paper [skip check-netlogo] [skip render-manual]
