### 2026-07-11 2d20e2d
docs: update documentation [skip check-netlogo] [skip render-manual]
