+++ index.qmd
@@ -27,47 +27,6 @@ If something is still unclear after reading through, feel free to open a discuss
 `LogoClim`'s interface. Click on the image to enlarge it.
 :::
 
-## Citation
-
-```{r}
-#| echo: false
-#| output: asis
-
-cat(
-  badge_custom(
-    x = "doi",
-    y = "10.17605/OSF.IO/EAPZU",
-    color = "1284C5",
-    url = "https://doi.org/10.17605/OSF.IO/EAPZU",
-    alt = ""
-  )
-)
-```
-
-::: {.callout-note}
-When using WorldClim data, you must also cite the original data sources.
-
-The appropriate citation depends on the specific dataset utilized. Please refer to the [WorldClim website](https://www.worldclim.org/data/index.html#citation) for up-to-date citation guidelines and dataset references.
-:::
-
-If you use this model in your research, please cite it to acknowledge the effort invested in its development and maintenance. Your citation helps support the ongoing improvement of the model.
-
-To cite `LogoClim` in publications please use the following format:
-
-Vartanian, D., Garcia, L., & Carvalho, A. M. (2026). *LogoClim: WorldClim in NetLogo* [Computer software]. <https://doi.org/10.17605/OSF.IO/EAPZU>
-
-A BibLaTeX entry for LaTeX users is:
-
-```latex
-@software{vartanian2026,
-  title = {LogoClim: WorldClim in NetLogo},
-  author = {{Daniel Vartanian} and {Leandro Garcia} and {Aline Martins de Carvalho}},
-  year = {2026},
-  publisher = {OSF},
-  doi = {10.17605/OSF.IO/EAPZU}
-}
-```
-
 ## Contributing
 
 ```{r}
@@ -85,11 +44,37 @@ cat(
 )
 ```
 
-Contributions are always welcome! Whether you want to report bugs, suggest new features, or help improve the code or documentation (including this manual), your input makes a difference.
+Contributions are always welcome, whether that's reporting bugs, suggesting features, or improving the code or docs.
 
 Before opening a new issue, please check the model's [issues tab](https://github.com/sustentarea/logoclim/issues) to see if your topic has already been reported.
 
 You can also support the development of `LogoClim` by becoming a sponsor. Click [here](https://github.com/sponsors/danielvartan) to make a donation. Please mention `LogoClim` in your donation message.
+
+## Citation
+
+::: {.callout-note}
+When using WorldClim data, you must also cite the original data sources.
+
+The appropriate citation depends on the specific dataset utilized. Please refer to the [WorldClim website](https://www.worldclim.org/data/index.html#citation) for up-to-date citation guidelines and dataset references.
+:::
+
+If you use this model in your research, please cite it to acknowledge the effort invested in its development and maintenance. Your citation helps support the ongoing improvement of the model.
+
+To cite `LogoClim` in publications please use the following format:
+
+Vartanian, D., Garcia, L., & Carvalho, A. M. (2026). *LogoClim: WorldClim in NetLogo* [Computer software]. GitHub. <https://github.com/danielvartan/logoclim>
+
+A BibLaTeX entry for LaTeX users is:
+
+```latex
+@software{vartanian2026,
+  title = {LogoClim: WorldClim in NetLogo},
+  author = {{Daniel Vartanian} and {Leandro Garcia} and {Aline Martins de Carvalho}},
+  year = {2026},
+  publisher = {GitHub},
+  url = {https://github.com/danielvartan/logoclim}
+}
+```
 
 ## License
 
@@ -119,13 +104,12 @@ We also acknowledge the World Climate Research Programme ([WCRP](https://www.wcr
 
[…]
