### 2026-07-11 e8a4ef5
docs: update documentation [skip check-netlogo] [skip render-manual]
