+++ paper/paper.bib
@@ -177,11 +177,11 @@
   keywords = {computer science,open science,r (programming language),ropensci,science,software reviews,sustainable development}
 }
 
-@software{vartanian2025g,
+@software{vartanian2026g,
   title = {Logônia: Plant growth response model in NetLogo},
   shorttitle = {Logônia},
   author = {Vartanian, Daniel and Garcia, Leandro and Carvalho, Aline Martins},
-  date = {2025},
+  date = {2026},
   url = {https://github.com/sustentarea/logonia},
   abstract = {Logônia is a NetLogo model that simulates the growth response of a fictional plant, logônia, under different climatic conditions. The model uses climate data from WorldClim 2.1 and demonstrates how to integrate the LogoClim model through the LevelSpace extension.},
   organization = {GitHub},
