+++ .github/workflows/check-netlogo.yaml
@@ -1,6 +1,7 @@
 on:
   push:
-    branches: [main, master]
+    paths:
+      - nlogox/**
 
 name: check-netlogo.yaml
 permissions: read-all
