### 2026-07-12 e460f81
feat: add headless checks when using `user-message`
