### 2026-07-12 2abfd3a
feat: add timeout to `check-netlogo.yaml` workflow
