+++ .github/workflows/draft-pdf.yml
@@ -12,7 +12,7 @@ jobs:
     name: Render Draft Paper
     steps:
       - name: Checkout
-        uses: actions/checkout@v4
+        uses: actions/checkout@v6
       - name: Build Draft PDF
         uses: openjournals/openjournals-draft-action@master
         with:
@@ -20,7 +20,7 @@ jobs:
           # This should be the path to the paper within your repo.
           paper-path: paper/paper.md
       - name: Upload
-        uses: actions/upload-artifact@v4
+        uses: actions/upload-artifact@v7
         with:
           name: paper
           # This is the output path where Pandoc will write the compiled
