+++ .github/workflows/check-netlogo.yaml
@@ -25,11 +25,11 @@ jobs:
     permissions: read-all
     steps:
       - name: Check out repository
-        uses: actions/checkout@v4
+        uses: actions/checkout@v6
 
       - name: Restore LogoClim test data cache
         id: logoclim-data-cache
-        uses: actions/cache@v4
+        uses: actions/cache@v6
         with:
           path: "data"
           key: "logoclim-data"
