### 2026-07-12 975bd43
fix: revert `headless?` checks when using `user-message`
