+++ index.qmd
@@ -44,7 +44,7 @@ cat(
 )
 ```
 
-Contributions are always welcome, whether that's reporting bugs, suggesting features, or improving the code or docs.
+Contributions are always welcome, whether that's reporting bugs, suggesting features, or improving the code or documentation.
 
 Before opening a new issue, please check the model's [issues tab](https://github.com/sustentarea/logoclim/issues) to see if your topic has already been reported.
 
