+++ qmd/data-processing.qmd
@@ -13,7 +13,7 @@ here("R", ".setup.R") |> source()
 ```
 
 ::: {.callout-note}
-`LogoClim` integrates [WorldClim](https://www.worldclim.org/) data into [NetLogo](https://www.netlogo.org/) models through a multi-step conversion process that can introduce minor discrepancies between the original data and the values used in the model. We recommend verifying data accuracy by running the [Near-Equality Tests](#sec-sm-near-equality-tests) notebook before proceeding with your analysis.
+`LogoClim` integrates [WorldClim](https://www.worldclim.org/) data into [NetLogo](https://www.netlogo.org/) models through a multi-step conversion process that can introduce minor discrepancies between the original data and the values used in the model. We recommend verifying data accuracy by running the [Near-Equality Tests](near-equality-tests.qmd) notebook before proceeding with your analysis.
 :::
 
 To convert the downloaded [GeoTIFF](https://en.wikipedia.org/wiki/GeoTIFF) files to [Esri ASCII](https://en.wikipedia.org/wiki/Esri_grid) grid format, we developed the [`worldclim_to_ascii()`](https://danielvartan.github.io/orbis/reference/worldclim_to_ascii.html) function, also part of the [`orbis`](https://danielvartan.github.io/orbis/index.html) R package [@vartanian2026a]. Just like [`worldclim_download()`](https://danielvartan.github.io/orbis/reference/worldclim_download.html), it is pretty straightforward to use. Here is a overview of its main parameters:
@@ -57,7 +57,7 @@ For example, if you want to crop [WorldClim](https://www.worldclim.org/) data to
 country_shape <- gadm("BRA", path = "data", level = 0)
 ```
 
-This code will download the shape of Brazil and store it in the `country_shape` variable, which can then be passed to the `shape` parameter of the [`worldclim_to_ascii`](https://danielvartan.github.io/orbis/reference/worldclim_to_ascii.html) function for cropping.
+This code will download the shape of Brazil and store it in the `country_shape` variable, which can then be passed to the `shape` parameter of the [`worldclim_to_ascii()`](https://danielvartan.github.io/orbis/reference/worldclim_to_ascii.html) function for cropping.
 
 ```{r}
 #| label: Plot country shape
@@ -131,7 +131,7 @@ Use the [`bbox finder`](https://bboxfinder.com) tool to easily obtain the coordi
 
 ## Conversion
 
-With the files downloaded and the shape or box defined, you can proceed to convert the data using the [`worldclim_to_ascii`](https://danielvartan.github.io/orbis/reference/worldclim_to_ascii.html) function. This function will take care of the conversion and cropping (if specified) in one step.
+With the files downloaded and the shape or box defined, you can proceed to convert the data using the [`worldclim_to_ascii()`](https://danielvartan.github.io/orbis/reference/worldclim_to_ascii.html) function. This function will take care of the conversion and cropping (if specified) in one step.
 
 For example, to work with average maximum temperature from the *Historical Climate Data* series at 10-minute resolution for Brazil, you could use the following code:
 
@@ -168,4 +168,4 @@ The function default values already takes care of possible distortions caused by
 
 After processing the data, you're ready to use `LogoClim`. Just open the model and point at your data directory using the model interface.
 
-For more details on available conversion parameters, refer to the [`worldclim_to_ascii`](https://danielvartan.github.io/orbis/reference/worldclim_to_ascii.html) documentation.
+For more details on available conversion parameters, refer to the [`worldclim_to_ascii()`](https://danielvartan.github.io/orbis/reference/worldclim_to_ascii.html) documentation.
