+++ qmd/near-equality-tests.qmd
@@ -17,7 +17,7 @@ This pipeline requires access to [WorldClim](https://www.worldclim.org/) and [GA
 
 ## Problem
 
-`LogoClim` integrates [WorldClim](https://www.worldclim.org/) data into [NetLogo](https://www.netlogo.org/) models through a multi-step conversion process. WorldClim provides raster data in [GeoTIFF](https://en.wikipedia.org/wiki/GeoTIFF) format, a high-precision binary format for geospatial data, but NetLogo's [GIS extension](https://docs.netlogo.org/gis.html) requires data in [Esri ASCII](https://en.wikipedia.org/wiki/Esri_grid) format, a text-based format that stores values with limited decimal precision. This conversion can introduce numerical differences through two main sources:
+`LogoClim` integrates [WorldClim](https://www.worldclim.org/) data into [NetLogo](https://www.netlogo.org/) models through a multi-step conversion process. WorldClim provides raster data in [GeoTIFF](https://en.wikipedia.org/wiki/GeoTIFF) format, a high-precision binary format for geospatial data, but NetLogo's [GIS extension](https://docs.netlogo.org/gis.html) requires data in [Esri ASCII](https://en.wikipedia.org/wiki/Esri_grid) format, a text-based format that stores values with a limited decimal precision. This conversion can introduce numerical differences through two main sources:
 
 1. **Decimal precision loss**: Esri ASCII files store values as text with a fixed number of decimal places, rounding continuous values from the original GeoTIFF.
 2. **Spatial interpolation**: When NetLogo's patch grid doesn't perfectly align with the raster resolution, the GIS extension applies interpolation that may alter cell values.
@@ -103,7 +103,7 @@ $$
 or
 
 $$
-\text{tolerance} \geq \frac{|x - y|}{\max(|x|, |y|)}
+\frac{|x - y|}{\max(|x|, |y|)} \leq \text{tolerance}
 $$
 
 where:
