+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 # Changelog
 
-All notable changes to this project will be documented in this file.
+All notable changes to this project will be documented below.
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+The format is based on [Keep a Changelog](https://keepachangelog.com/). The project also follows [Semantic Versioning](https://semver.org/) and [Conventional Commits](https://www.conventionalcommits.org/).
 
-## v2.2.0.9000 (development version)
+## v2.2.0 (2026-07-13)
 
 ### Added
 
