+++ CITATION.cff
@@ -55,4 +55,4 @@ identifiers:
 repository-code: https://github.com/sustentarea/logoclim
 repository-artifact: https://doi.org/10.17605/OSF.IO/RE95Z
 license: GPL-3.0-or-later
-version: 2.2.0.9000
+version: 2.2.0
