### 2026-07-13 63e0297
release: prepare for v2.2.1 release
