+++ paper/paper.bib
@@ -182,8 +182,8 @@
   shorttitle = {Logônia},
   author = {Vartanian, Daniel and Garcia, Leandro and Carvalho, Aline Martins},
   date = {2026},
-  url = {https://github.com/sustentarea/logonia},
-  abstract = {Logônia is a NetLogo model that simulates the growth response of a fictional plant, logônia, under different climatic conditions. The model uses climate data from WorldClim 2.1 and demonstrates how to integrate the LogoClim model through the LevelSpace extension.},
-  organization = {GitHub},
+  doi = {10.5281/zenodo.21332539},
+  abstract = {Logônia is a NetLogo model that simulates the growth response of a fictional plant, Logônia, under different climatic conditions. The model uses climate data from WorldClim 2.1 and demonstrates how to integrate the LogoClim model through the LevelSpace extension.},
+  organization = {Zenodo},
   keywords = {climate change,computer science,environmental sciences,exact sciences,interdisciplinary fields,logoclim,netlogo,nosource,probability and statistics,worldclim}
 }
