+++ qmd/how-it-works.qmd
@@ -19,6 +19,6 @@ here("R", ".setup.R") |> source()
 
 The model operates on a grid of patches, where each patch represents a geographical area and stores values for latitude, longitude, and selected climate variables. During the simulation, patches update their colors based on the data values. The results can be visualized on a map, accompanied by plots that display the mean, minimum, maximum, and standard deviation of the selected variable over time.
 
-All climate inputs come from [WorldClim 2.1](https://worldclim.org/), a widely used source of high-resolution climate datasets [@fick2017]. These data series are offered at various spatial resolutions, ranging from 10 minutes (~340 km² at the equator) to 30 seconds (~1 km² at the equator), and can be chosen within the model interface.
+All climate inputs come from [WorldClim 2.1](https://worldclim.org/), a widely used source of high-resolution climate datasets [@fick2017]. These data series are offered at various spatial resolutions, ranging from 10 minutes (about 340 km² at the equator) to 30 seconds (about 1 km² at the equator).
 
 Learn more about the data series in the [WorldClim](worldclim.qmd) section.
