+++ references.bib
@@ -224,6 +224,17 @@
   keywords = {agent-based models,automation,computer science,continuous integration,exact sciences,interdisciplinary fields,netlogo,netlogo models,nosource,probability and statistics,simulations,unit tests}
 }
 
+@software{vartanian2026g,
+  title = {Logônia: Plant growth response model in NetLogo},
+  shorttitle = {Logônia},
+  author = {Vartanian, Daniel and Garcia, Leandro and Carvalho, Aline Martins},
+  date = {2026},
+  doi = {10.5281/zenodo.21332539},
+  abstract = {Logônia is a NetLogo model that simulates the growth response of a fictional plant, Logônia, under different climatic conditions. The model uses climate data from WorldClim 2.1 and demonstrates how to integrate the LogoClim model through the LevelSpace extension.},
+  organization = {Zenodo},
+  keywords = {climate change,computer science,environmental sciences,exact sciences,interdisciplinary fields,logoclim,netlogo,nosource,probability and statistics,worldclim}
+}
+
 @article{wickham2011,
   title = {{{testthat}}: Get started with testing},
   shorttitle = {{{testthat}}},
