### 2026-07-13 8c2407a
fix: add `ISL` exception in near-equality-tests.qmd
