### 2026-07-13 5e4ed4d
docs: bump CoMSES Network badge to 2.2.1 [skip render-manual]
