### 2026-07-13 8fe0648
docs: fix citation and DOI links to Zenodo in the user manual
