### 2026-07-13 d843e4e
docs: update LevelSpace links to new NetLogo documentation
