+++ qmd/how-to-use-it.qmd
@@ -90,7 +90,7 @@ Here is a brief overview of the model interface and its controls.
 
 ## Integrating with Other Models
 
-`LogoClim` was created to be integrated with other models using NetLogo's [LevelSpace](https://ccl.northwestern.edu/netlogo/docs/ls.html) extension. This extension enables parallel execution and data exchange between models. The following chapters of this manual provide guidance on how to integrate `LogoClim` with other model.
+`LogoClim` was created to be integrated with other models using NetLogo's [LevelSpace](https://docs.netlogo.org/ls) extension. This extension enables parallel execution and data exchange between models. The following chapters of this manual provide guidance on how to integrate `LogoClim` with other model.
 
 To facilitate this integration, we created the [`Logônia`](https://github.com/sustentarea/logonia) model [@vartanian2026g], a fictional plant-growth model providing a practical example of how to integrate `LogoClim`. It is also available on the [CoMSES Network](https://www.comses.net/codebases/4f2be13a-3957-4537-bf64-3fad96ba271f/) and its code repository is available on [GitHub](https://github.com/sustentarea/logonia).
 
