### 2026-07-18 0b0944c
docs: update paper.md with minor text edits and formatting changes
