### 2026-08-18 #8845 @kthyng
Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file. For example "brazil".
