### 2026-08-18 89797f0
docs: fix capitalization in paper references
