+++ paper/paper.bib
@@ -47,7 +47,7 @@
 }
 
 @dataset{carvalho2023a,
-  title = {Global syndemic: The impact of anthropogenic climate change on the health and nutrition of children under five years old attended by Brazil's public health system (SUS)},
+  title = {Global syndemic: The impact of anthropogenic climate change on the health and nutrition of children under five years old attended by {Brazil}'s public health system (SUS)},
   author = {Carvalho, Aline Martins and family=Albuquerque, given=Ulysses Paulino, prefix=de, useprefix=false and Barbosa, Brena Barreto and Magalhães, Arthur Ramalho and Klapka, Cecilia Stanzani and Vartanian, Daniel and Pereira, Estela Barici and Schettino, João Pedro Junqueira},
   date = {2023},
   publisher = {OSF},
